Annotate an EML file (.xml) or an emld list object

Annotate an EML file (.xml) or an emld list object

annotate_eml(annotations = NULL, eml.in = NULL, eml.out = NULL)

Arguments

annotations

(character; data frame) Path or data frame of the annotations.txt template for annotating eml.in. Create annotations.txt with template_annotations().

eml.in

(EML; emld list) Path to, or emld list of, the EML file to be annotated. Accepted versions of the emld list object can be created by EMLassemblyline::read_eml() or make_eml().

eml.out

(character) Path of the annotated EML file to be written.

Value

(EML; emld list) If the input to eml.in is an EML file then the output will be an EML file. The emld list object is always returned.

Note

All annotated elements are assigned ids and their annotations are placed both immediately under the parent element (subject) when supported and within the /eml/annotations node through id+reference pairs. This redundant approach supports variation in where EML metadata consumers harvest this information and supports annotation of EML elements requiring id+references.

Examples

# Annotate an EML file --------------------------------------------------------
#
# Annotate an existing EML file (edi.260.3.xml) and write a new version 
# (edi.260.4.xml)

eml <- annotate_eml(
 annotations = system.file("/examples/pkg_260/metadata_templates_overflow/annotations.txt", package = "EMLassemblyline"),
 eml.in = system.file("/examples/eml/edi.260.3.xml", package = "EMLassemblyline"),
 eml.out = paste0(tempdir(), "/edi.260.4.xml")
)
#> Annotating EML ...
#> Checking inputs
#> Done.
eml
#> access:
#>   authSystem: https://pasta.edirepository.org/authentication
#>   order: allowFirst
#>   scope: document
#>   allow:
#>   - principal: uid=userid1,o=EDI,dc=edirepository,dc=org
#>     permission: all
#>   - principal: uid=userid2,o=LTER,dc=ecoinformatics,dc=org
#>     permission: all
#>   - principal: public
#>     permission: read
#> additionalMetadata:
#>   metadata:
#>     unitList:
#>       unit:
#>       - id: gramsPerSquaredMeterPerCentimeter
#>         multiplierToSI: '0.1'
#>         name: gramsPerSquaredMeterPerCentimeter
#>         parentSI: kilogramsPerSquaredMeterPerMeter
#>         unitType: areal mass density per length
#>         description: grams per meter squared per centimeter
#>       - id: milligramsPerGram
#>         multiplierToSI: '1000'
#>         name: milligramsPerGram
#>         parentSI: kilogramsPerSquaredMeterPerMeter
#>         unitType: mass
#>         description: milligrams per gram
#> dataset:
#>   title: 'Sphagnum and Vascular Plant Decomposition under Increasing Nitrogen Additions:
#>     2014-2015'
#>   pubDate: '2020-10-11'
#>   abstract:
#>     para: |
#>       Development of the oil sands has led to increasing atmospheric N deposition, with values as high as 17 kg N ha^-1 yr^-1; regional background levels &lt;2 kg N ha^-1 yr^-1.  Bogs, being ombrotrophic, may be especially susceptible to increasing N deposition.  To examine responses to N deposition, over five years, we experimentally applied N (as NH4NO3) to a bog near Mariana Lakes, Alberta, at rates of 0, 5, 10, 15, 20, and 25 kg N ha^-1 yr^-1, plus controls (no water or N addition).   In May of each year, we collected mixed vascular plant tissue and Sphagnum fuscum peat and placed homogenized mixtures in nylon bags and placed them approximately 10 cm below the peat surface in early June.  Bags were collected again in October of each year, cleaned, dried, and weighed.  Decomposition of Sphagnum moss and mixed vascular plant litter was affected by N inputs, on average losing 8 and 38 % of initial mass, respectively, over 5 months of decomposition. Water addition alone had no significant effect on decomposition of cellulose, Sphagnum, or vascular plant litter (p &gt; 0.15).&#13;
#>   additionalInfo:
#>     para: 'Articles derived from this dataset: Article 1, Article 2, Article 3'
#>   intellectualRights:
#>     para: |
#>       This data package is released to the \"public domain\" under Creative Commons CC0 1.0 \"No Rights Reserved\" (see: https://creativecommons.org/publicdomain/zero/1.0/). It is considered professional etiquette to provide attribution of the original work if this data package is shared in whole or by individual components. A generic citation is provided for this data package on the website https://portal.edirepository.org (herein \"website\") in the summary metadata page. Communication (and collaboration) with the creators of this data package is recommended to prevent duplicate research or publication. This data package (and its components) is made available \"as is\" and with no warranty of accuracy or fitness for use. The creators of this data package and the website shall not be liable for any damages resulting from misinterpretation or misuse of the data package or its components. Periodic updates of this data package may be available from the website. Thank you.&#13;
#>   coverage:
#>     temporalCoverage:
#>       rangeOfDates:
#>         beginDate:
#>           calendarDate: '2014-05-01'
#>         endDate:
#>           calendarDate: '2015-10-31'
#>     geographicCoverage:
#>     - geographicDescription: Alberta, Canada, 100 km south of Fort McMurray, Canada
#>       boundingCoordinates:
#>         westBoundingCoordinate: '112.094'
#>         eastBoundingCoordinate: '112.094'
#>         northBoundingCoordinate: '55.895'
#>         southBoundingCoordinate: '55.895'
#>     - geographicDescription: site_1
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.43'
#>         eastBoundingCoordinate: '-89.43'
#>         northBoundingCoordinate: '45.23'
#>         southBoundingCoordinate: '45.23'
#>     - geographicDescription: site_2
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.44'
#>         eastBoundingCoordinate: '-89.44'
#>         northBoundingCoordinate: '45.24'
#>         southBoundingCoordinate: '45.24'
#>     - geographicDescription: site_3
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.45'
#>         eastBoundingCoordinate: '-89.45'
#>         northBoundingCoordinate: '45.25'
#>         southBoundingCoordinate: '45.25'
#>     - geographicDescription: site_4
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.46'
#>         eastBoundingCoordinate: '-89.46'
#>         northBoundingCoordinate: '45.26'
#>         southBoundingCoordinate: '45.26'
#>     - geographicDescription: site_5
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.47'
#>         eastBoundingCoordinate: '-89.47'
#>         northBoundingCoordinate: '45.27'
#>         southBoundingCoordinate: '45.27'
#>     - geographicDescription: site_6
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.48'
#>         eastBoundingCoordinate: '-89.48'
#>         northBoundingCoordinate: '45.28'
#>         southBoundingCoordinate: '45.28'
#>     - geographicDescription: site_7
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.49'
#>         eastBoundingCoordinate: '-89.49'
#>         northBoundingCoordinate: '45.29'
#>         southBoundingCoordinate: '45.29'
#>     - geographicDescription: site_8
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.5'
#>         eastBoundingCoordinate: '-89.5'
#>         northBoundingCoordinate: '45.3'
#>         southBoundingCoordinate: '45.3'
#>     - geographicDescription: site_9
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.51'
#>         eastBoundingCoordinate: '-89.51'
#>         northBoundingCoordinate: '45.31'
#>         southBoundingCoordinate: '45.31'
#>     - geographicDescription: site_10
#>       boundingCoordinates:
#>         westBoundingCoordinate: '-89.52'
#>         eastBoundingCoordinate: '-89.52'
#>         northBoundingCoordinate: '45.32'
#>         southBoundingCoordinate: '45.32'
#>   maintenance:
#>     description: Completed
#>   methods:
#>     methodStep:
#>       description:
#>         para: In 2014 and 2015, we quantified decomposition of plant litter material
#>           collected from each plot.  In late May, we collected vascular plant litter
#>           from the surface of each plot, along with <emphasis>Sphagnum</emphasis>
#>           peat (<emphasis>S. fuscum</emphasis>-derived), from roughly 5-15 cm below
#>           the capitula.  We placed nylon mesh litter bags with mixed vascular plant
#>           litter (approx. 1.7 g dry mass; roughly equal masses of <emphasis>Rhododendron
#>           groenlandicum</emphasis>, <emphasis>Andromeda polifolia</emphasis>, <emphasis>Chamaedaphne
#>           calyculata</emphasis>, <emphasis>Smilacina trifolia</emphasis>, and <emphasis>Rubus
#>           chamaemorus</emphasis>) or peat approximately 10 cm below the peat surface
#>           in early June (in 2014, 3 bags per plot of vascular litter and 3 bags of
#>           <emphasis>Sphagnum</emphasis> peat; in 2015, 4 bags of each material).  Upon
#>           retrieval in October of each year, we manually removed debris from each
#>           retrieved bag before drying at 55 C for a minimum of 5 days, and weighing.
#>   project:
#>     title: 'Nitrogen Addition Experiments in Boreal Ecosystems: Understanding the
#>       Fate of Atmospherically Deposited Nitrogen in order to Determine Nitrogen Critical
#>       Loads'
#>     personnel:
#>     - - individualName:
#>           surName: Wieder
#>           givenName:
#>           - Kelman
#>           - R
#>         organizationName: Villanova University
#>         electronicMailAddress: kelman.wieder@villanova.edu
#>         userId:
#>           directory: https://orcid.org
#>           userId: https://orcid.org/0000-0001-5856-8473
#>         role: Principal Investigator
#>     funding: Cumulative Environmental Management Association (CEMA) 2011-0003
#>     relatedProject:
#>     - title: Decomposition in the real world
#>       personnel:
#>       - individualName:
#>           surName: Vile
#>           givenName:
#>           - Melanie
#>           - A
#>         organizationName: Villanova University
#>         electronicMailAddress: mvile@villanova.edu
#>         role: Principal Investigator
#>       funding: NSF 111001
#>     - title: Project title 3
#>       personnel:
#>       - individualName:
#>           givenName: Joe
#>           surName: Smith
#>         organizationName: Villanova University
#>         role: Principal Investigator
#>       funding: NSF 101
#>     - title: Project title 4
#>       personnel:
#>       - individualName:
#>           givenName: Chase
#>           surName: Gacho
#>         organizationName: University of New Mexico
#>         electronicMailAddress: gachase@unm.edu
#>         role: Principal Investigator
#>       funding: NSF 111001
#>   creator:
#>   - individualName:
#>       surName: Wieder
#>       givenName:
#>       - Kelman
#>       - R
#>     organizationName: Villanova University
#>     electronicMailAddress: kelman.wieder@villanova.edu
#>     userId:
#>       directory: https://orcid.org
#>       userId: https://orcid.org/0000-0001-5856-8473
#>     id: /Kelman R Wieder e9f6fad4-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       surName: Vitt
#>       givenName:
#>       - Dale
#>       - H
#>     organizationName: Southern Illinois University
#>     id: /Dale H Vitt e9f713f2-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       surName: Vile
#>       givenName:
#>       - Melanie
#>       - A
#>     organizationName: Villanova University
#>     electronicMailAddress: mvile@villanova.edu
#>     id: /Melanie A Vile e9f729b4-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       surName: Graham
#>       givenName:
#>       - Jeremy
#>       - A
#>     organizationName: Southern Illinois University
#>     electronicMailAddress: jeremyg@mtu.edu
#>     id: /Jeremy A Graham e9f7402a-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       surName: Hartsock
#>       givenName:
#>       - Jeremy
#>       - A
#>     organizationName: Southern Illinois University
#>     electronicMailAddress: hartsock@siu.edu
#>     id: /Jeremy A Hartsock e9f75510-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       givenName: Hope
#>       surName: Fillingim
#>     organizationName: Villanova University
#>     electronicMailAddress: fillingimhope@gmail.com
#>     id: /Hope Fillingim e9f76ab4-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       givenName: Melissa
#>       surName: House
#>     organizationName: Southern Illinois University
#>     electronicMailAddress: mhouse@siu.edu
#>     id: /Melissa House e9f78012-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       surName: Quinn
#>       givenName:
#>       - James
#>       - C
#>     organizationName: Villanova University
#>     electronicMailAddress: jcqui03@gmail.com
#>     id: /James C Quinn e9f794a8-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       surName: Scott
#>       givenName:
#>       - Kimberli
#>       - D
#>     organizationName: Villanova University
#>     electronicMailAddress: kscott@villanova.edu
#>     id: /Kimberli D Scott e9f7aa4c-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       givenName: Meaghan
#>       surName: Petix
#>     organizationName: Southern Illinois University
#>     electronicMailAddress: meaghan.petix@wsu.edu
#>     id: /Meaghan Petix e9f7be9c-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       surName: McMillen
#>       givenName:
#>       - Kelly
#>       - J
#>     organizationName: Villanova University
#>     electronicMailAddress: Kelly.Mcmillen@ttu.edu
#>     id: /Kelly J McMillen e9f7d3f0-fa9f-11ef-acd1-0022484ed5a5
#>   associatedParty:
#>   - organizationName: Center for Limnology
#>     electronicMailAddress: cfl@wisc.edu
#>     userId:
#>       directory: https://orcid.org
#>       userId: https://orcid.org/0000-0001-5856-8473
#>     role: Primary Collaborator
#>   - organizationName: Trout Lake Station
#>     electronicMailAddress: tls@wisc.edu
#>     role: Secondary Collaborator
#>   - individualName:
#>       surName: Quinn
#>       givenName:
#>       - James
#>       - C
#>     organizationName: Villanova University
#>     electronicMailAddress: jcqui03@gmail.com
#>     role: Instrument Technician
#>     id: /James C Quinn e9f80028-fa9f-11ef-acd1-0022484ed5a5
#>   - individualName:
#>       givenName: Jane
#>       surName: Doe
#>     organizationName: Center for Limnology
#>     electronicMailAddress: jdoe@email.com
#>     userId:
#>       directory: https://orcid.org
#>       userId: https://orcid.org/0000-0001-5856-8473
#>     role: Equipment Lead
#>     id: /Jane Doe e9f81572-fa9f-11ef-acd1-0022484ed5a5
#>   keywordSet:
#>   - keyword:
#>     - Villanova Peatland Biogeochemistry Group
#>     - peat
#>     - vascular
#>     - moss
#>   - keywordThesaurus: LTER Controlled Vocabulary
#>     keyword:
#>     - bog
#>     - decomposition
#>   contact:
#>   - individualName:
#>       surName: Wieder
#>       givenName:
#>       - Kelman
#>       - R
#>     organizationName: Villanova University
#>     electronicMailAddress: kelman.wieder@villanova.edu
#>     userId:
#>       directory: https://orcid.org
#>       userId: https://orcid.org/0000-0001-5856-8473
#>     id: /Kelman R Wieder e9f7e886-fa9f-11ef-acd1-0022484ed5a5
#>   - organizationName: Villanova University
#>     positionName: Information Manager
#>     electronicMailAddress: info@villanova.edu
#>     userId:
#>       directory: https://orcid.org
#>       userId: https://orcid.org/0000-0001-5856-8473
#>   - organizationName: Villanova University
#>     positionName: Organization Lead
#>     electronicMailAddress: info@villanova.edu
#>   dataTable:
#>   - entityName: Decomp file name
#>     entityDescription: Decomp file description
#>     physical:
#>       objectName: decomp.csv
#>       size:
#>         unit: bytes
#>         size: '15431'
#>       authentication:
#>         method: MD5
#>         authentication: 90f84458e577ba57c0204dc5a32030dd
#>       dataFormat:
#>         textFormat:
#>           numHeaderLines: '1'
#>           recordDelimiter: \r\n
#>           attributeOrientation: column
#>           simpleDelimited:
#>             fieldDelimiter: ','
#>             quoteCharacter: \'
#>       distribution:
#>         online:
#>           url: https://url/to/decomp.csv
#>     attributeList:
#>       attribute:
#>       - attributeName: type
#>         attributeDefinition: 'Vegetation type: Sphagnum fuscum or vascular plant tissue'
#>         storageType: string
#>         measurementScale:
#>           nominal:
#>             nonNumericDomain:
#>               enumeratedDomain:
#>                 codeDefinition:
#>                 - code: Sphagnum
#>                   definition: Sphagnum
#>                 - code: Vascular
#>                   definition: Vascular
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /decomp.csv/type
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002623
#>             label: type of soil core
#>       - attributeName: date
#>         attributeDefinition: 'Year of growth: 2011-2015'
#>         storageType: date
#>         measurementScale:
#>           dateTime:
#>             formatString: YYYY-MM-DD
#>             dateTimeDomain: []
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /decomp.csv/date
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002043
#>             label: date and time of measurement
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002072
#>             label: year and day of measurement
#>       - attributeName: arm
#>         attributeDefinition: Experimental plots were located along 3 arms; each arm
#>           had one replicate each of 8 N treatments
#>         storageType: string
#>         measurementScale:
#>           nominal:
#>             nonNumericDomain:
#>               enumeratedDomain:
#>                 codeDefinition:
#>                 - code: '1'
#>                   definition: Arm 1
#>                 - code: '2'
#>                   definition: Arm 2
#>                 - code: '3'
#>                   definition: Arm 3
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /decomp.csv/arm
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002997
#>             label: site identifier
#>       - attributeName: ntrt
#>         attributeDefinition: Field experimental N treatments were made with 8 applications
#>           per plot per year between mid-May and mid-August of each year
#>         storageType: string
#>         measurementScale:
#>           nominal:
#>             nonNumericDomain:
#>               enumeratedDomain:
#>                 codeDefinition:
#>                 - code: C
#>                   definition: Control, no water, no N added
#>                 - code: '0'
#>                   definition: 0 kg N/ha/yr added
#>                 - code: '5'
#>                   definition: 5 kg N/ha/yr added
#>                 - code: '10'
#>                   definition: 10 kg N/ha/yr added as NH4NO3
#>                 - code: '15'
#>                   definition: 15 kg N/ha/yr added
#>                 - code: '20'
#>                   definition: 20 kg N/ha/yr added as NH4NO3
#>                 - code: '25'
#>                   definition: 25 kg N/ha/yr added
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /decomp.csv/ntrt
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002216
#>             label: nitrate concentration
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0010049
#>             label: gram per square meter
#>       - attributeName: year
#>         attributeDefinition: Year collected
#>         storageType: date
#>         measurementScale:
#>           dateTime:
#>             formatString: YYYY
#>             dateTimeDomain: []
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /decomp.csv/year
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002050
#>             label: year of measurement
#>       - attributeName: percent_loss
#>         attributeDefinition: Amount of mass lost expressed as a percent of original
#>           mass
#>         storageType: float
#>         measurementScale:
#>           ratio:
#>             unit:
#>               standardUnit: dimensionless
#>             numericDomain:
#>               numberType: real
#>               bounds:
#>                 minimum:
#>                   exclusive: 'false'
#>                   minimum: '0'
#>                 maximum:
#>                   exclusive: 'false'
#>                   maximum: '57.65'
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /decomp.csv/percent_loss
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002294
#>             label: loss on ignition percentage
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0000187
#>             label: percent
#>       - attributeName: taxa
#>         attributeDefinition: Taxa observed at site
#>         storageType: string
#>         measurementScale:
#>           nominal:
#>             nonNumericDomain:
#>               textDomain:
#>                 definition: Taxa observed at site
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /decomp.csv/taxa
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002730
#>             label: taxonomic classification
#>     numberOfRecords: '294'
#>     id: /decomp.csv
#>     annotation:
#>     - propertyURI:
#>         '': http://purl.obolibrary.org/obo/IAO_0000136
#>         label: is about
#>       valueURI:
#>         '': http://purl.dataone.org/odo/ECSO_00000034
#>         label: Leaf Litter Carbon Pool
#>   - entityName: Nitrogen file name
#>     entityDescription: Nitrogen file description
#>     physical:
#>       objectName: nitrogen.csv
#>       size:
#>         unit: bytes
#>         size: '6297'
#>       authentication:
#>         method: MD5
#>         authentication: e6609e09690640fb64b104fd5e8b6d4e
#>       dataFormat:
#>         textFormat:
#>           numHeaderLines: '1'
#>           recordDelimiter: \r
#>           attributeOrientation: column
#>           simpleDelimited:
#>             fieldDelimiter: ','
#>             quoteCharacter: \'
#>       distribution:
#>         online:
#>           url: https://url/to/nitrogen.csv
#>     attributeList:
#>       attribute:
#>       - attributeName: date
#>         attributeDefinition: 'Year of growth: 2011-2015'
#>         storageType: date
#>         measurementScale:
#>           dateTime:
#>             formatString: YYYY-MM-DD
#>             dateTimeDomain: []
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/date
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002043
#>             label: date and time of measurement
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002072
#>             label: year and day of measurement
#>       - attributeName: ntrt
#>         attributeDefinition: Field experimental N treatments were made with 8 applications
#>           per plot per year between mid May and mid August of each year
#>         storageType: string
#>         measurementScale:
#>           nominal:
#>             nonNumericDomain:
#>               enumeratedDomain:
#>                 codeDefinition:
#>                 - code: C
#>                   definition: Control, no water, no N added
#>                 - code: '0'
#>                   definition: 0 kg N/ha/yr added as NH4NO3
#>                 - code: '5'
#>                   definition: 5 kg N/ha/yr added as NH4NO3
#>                 - code: '10'
#>                   definition: 10 kg N/ha/yr added as NH4NO3
#>                 - code: '15'
#>                   definition: 15 kg N/ha/yr added as NH4NO3
#>                 - code: '20'
#>                   definition: 20 kg N/ha/yr added as NH4NO3
#>                 - code: '25'
#>                   definition: 25 kg N/ha/yr added as NH4NO3
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/ntrt
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002216
#>             label: nitrate concentration
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0010049
#>             label: gram per square meter
#>       - attributeName: arm
#>         attributeDefinition: Experimental plots were located along 3 arms; each arm
#>           had one replicate each of 8 N treatments
#>         storageType: string
#>         measurementScale:
#>           nominal:
#>             nonNumericDomain:
#>               enumeratedDomain:
#>                 codeDefinition:
#>                 - code: '1'
#>                   definition: Arm 1
#>                 - code: '2'
#>                   definition: Arm 2
#>                 - code: '3'
#>                   definition: Arm3
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/arm
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002997
#>             label: site identifier
#>       - attributeName: net_primary_production
#>         attributeDefinition: The product of moss linear growth of stems (cm yr-1)
#>           and stem mass density (SMD)
#>         storageType: float
#>         measurementScale:
#>           ratio:
#>             unit:
#>               standardUnit: gramsPerMeterSquaredPerYear
#>             numericDomain:
#>               numberType: real
#>               bounds:
#>                 minimum:
#>                   exclusive: 'false'
#>                   minimum: '39.4'
#>                 maximum:
#>                   exclusive: 'false'
#>                   maximum: '540.7'
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/net_primary_production
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00001137
#>             label: Net Primary Production Biomass Flux
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0000311
#>             label: milligram per kilogram per day
#>       - attributeName: stem_mass_density
#>         attributeDefinition: Mass of 1-cm lengths of Sphagnum stems beneath the capitula
#>           per m2
#>         storageType: float
#>         measurementScale:
#>           ratio:
#>             unit:
#>               customUnit: gramsPerSquaredMeterPerCentimeter
#>             numericDomain:
#>               numberType: real
#>               bounds:
#>                 minimum:
#>                   exclusive: 'false'
#>                   minimum: '29.5'
#>                 maximum:
#>                   exclusive: 'false'
#>                   maximum: '268.8'
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/stem_mass_density
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00000678
#>             label: mass density
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0000182
#>             label: density unit
#>       - attributeName: plant_density
#>         attributeDefinition: Number of plants/area
#>         storageType: float
#>         measurementScale:
#>           ratio:
#>             unit:
#>               standardUnit: numberPerMeterSquared
#>             numericDomain:
#>               numberType: natural
#>               bounds:
#>                 minimum:
#>                   exclusive: 'false'
#>                   minimum: '20191'
#>                 maximum:
#>                   exclusive: 'false'
#>                   maximum: '87719'
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/plant_density
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00001147
#>             label: Plant Material Biomass Density
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0000052
#>             label: mass density unit
#>       - attributeName: stem_n_concentration
#>         attributeDefinition: N concentration in the 0-2 cm stem sections
#>         storageType: float
#>         measurementScale:
#>           ratio:
#>             unit:
#>               customUnit: milligramsPerGram
#>             numericDomain:
#>               numberType: real
#>               bounds:
#>                 minimum:
#>                   exclusive: 'false'
#>                   minimum: '6.83'
#>                 maximum:
#>                   exclusive: 'false'
#>                   maximum: '12.92'
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/stem_n_concentration
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00001883
#>             label: nitrogen concentration
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0000051
#>             label: concentration unit
#>       - attributeName: n_assimilated
#>         attributeDefinition: Amount of N used and stored per area of moss cover per
#>           year
#>         storageType: float
#>         measurementScale:
#>           ratio:
#>             unit:
#>               standardUnit: gramsPerMeterSquaredPerYear
#>             numericDomain:
#>               numberType: real
#>               bounds:
#>                 minimum:
#>                   exclusive: 'false'
#>                   minimum: '0.4'
#>                 maximum:
#>                   exclusive: 'false'
#>                   maximum: '5.26'
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/n_assimilated
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00001123
#>             label: http://purl.dataone.org/odo/ECSO_00001883
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0000311
#>             label: milligram per kilogram per day
#>       - attributeName: site_name
#>         attributeDefinition: Name of site
#>         storageType: string
#>         measurementScale:
#>           nominal:
#>             nonNumericDomain:
#>               textDomain:
#>                 definition: Name of site
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/site_name
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002392
#>             label: station name
#>       - attributeName: site_lat
#>         attributeDefinition: Latitude in decimal degrees
#>         storageType: float
#>         measurementScale:
#>           ratio:
#>             unit:
#>               standardUnit: degree
#>             numericDomain:
#>               numberType: real
#>               bounds:
#>                 minimum:
#>                   exclusive: 'false'
#>                   minimum: '45.23'
#>                 maximum:
#>                   exclusive: 'false'
#>                   maximum: '46.26'
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/site_lat
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002130
#>             label: latitude coordinate
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0000185
#>             label: degree
#>       - attributeName: site_lon
#>         attributeDefinition: Longitude in decimal degrees
#>         storageType: float
#>         measurementScale:
#>           ratio:
#>             unit:
#>               standardUnit: degree
#>             numericDomain:
#>               numberType: real
#>               bounds:
#>                 minimum:
#>                   exclusive: 'false'
#>                   minimum: '-90.46'
#>                 maximum:
#>                   exclusive: 'false'
#>                   maximum: '-89.43'
#>         missingValueCode:
#>           code: '-99999'
#>           codeExplanation: Missing value
#>         id: /nitrogen.csv/site_lon
#>         annotation:
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>             label: contains measurements of type
#>           valueURI:
#>             '': http://purl.dataone.org/odo/ECSO_00002132
#>             label: longitude coordinate
#>         - propertyURI:
#>             '': http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>             label: has unit
#>           valueURI:
#>             '': http://purl.obolibrary.org/obo/UO_0000185
#>             label: degree
#>     numberOfRecords: '104'
#>     id: /nitrogen.csv
#>     annotation:
#>     - propertyURI:
#>         '': http://purl.obolibrary.org/obo/IAO_0000136
#>         label: is about
#>       valueURI:
#>         '': http://purl.dataone.org/odo/ECSO_00002467
#>         label: nitrogen proportion in whole plant
#>   otherEntity:
#>   - entityName: ancillary_data file name
#>     entityDescription: ancillary_data file description
#>     physical:
#>       objectName: ancillary_data.zip
#>       size:
#>         unit: bytes
#>         size: '4189'
#>       authentication:
#>         method: MD5
#>         authentication: f128eaf557babfbb85efeaef488e9c94
#>       dataFormat:
#>         externallyDefinedFormat:
#>           formatName: application/zip
#>       distribution:
#>         online:
#>           url: https://url/to/ancillary_data.zip
#>     entityType: unknown
#>     id: /ancillary_data.zip
#>     annotation:
#>     - propertyURI:
#>         '': http://purl.obolibrary.org/obo/IAO_0000136
#>         label: is about
#>       valueURI:
#>         '': http://purl.obolibrary.org/obo/ENVO_00002297
#>         label: environmental feature
#>   - entityName: processing_and_analysis file name
#>     entityDescription: processing_and_analysis file description
#>     physical:
#>       objectName: processing_and_analysis.R
#>       size:
#>         unit: bytes
#>         size: '2230'
#>       authentication:
#>         method: MD5
#>         authentication: bb3e4aba767dfb2d4e3b053e1dfa1985
#>       dataFormat:
#>         externallyDefinedFormat:
#>           formatName: text/plain
#>       distribution:
#>         online:
#>           url: https://url/to/processing_and_analysis.R
#>     entityType: unknown
#>     id: /processing_and_analysis.R
#>     annotation:
#>     - propertyURI:
#>         '': http://purl.obolibrary.org/obo/IAO_0000136
#>         label: is about
#>       valueURI:
#>         '': http://purl.dataone.org/odo/ECSO_00002489
#>         label: analysis code
#>   id: /dataset
#>   annotation:
#>   - propertyURI:
#>       '': http://purl.obolibrary.org/obo/IAO_0000136
#>       label: is about
#>     valueURI:
#>       '': http://purl.dataone.org/odo/ECSO_00002467
#>       label: nitrogen proportion in whole plant
#> packageId: edi.260.4
#> schemaLocation: eml://ecoinformatics.org/eml-2.2.0  http://nis.lternet.edu/schemas/EML/eml-2.2.0/xsd/eml.xsd
#> system: edi
#> annotations:
#>   annotation:
#>   - references: dataset
#>     propertyURI:
#>       label: is about
#>       propertyURI: http://purl.obolibrary.org/obo/IAO_0000136
#>     valueURI:
#>       label: nitrogen proportion in whole plant
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002467
#>   - references: nitrogen.csv
#>     propertyURI:
#>       label: is about
#>       propertyURI: http://purl.obolibrary.org/obo/IAO_0000136
#>     valueURI:
#>       label: nitrogen proportion in whole plant
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002467
#>   - references: nitrogen.csv/date
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: date and time of measurement
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002043
#>   - references: nitrogen.csv/date
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: year and day of measurement
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002072
#>   - references: nitrogen.csv/ntrt
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: nitrate concentration
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002216
#>   - references: nitrogen.csv/ntrt
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: gram per square meter
#>       valueURI: http://purl.obolibrary.org/obo/UO_0010049
#>   - references: nitrogen.csv/arm
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: site identifier
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002997
#>   - references: nitrogen.csv/net_primary_production
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: Net Primary Production Biomass Flux
#>       valueURI: http://purl.dataone.org/odo/ECSO_00001137
#>   - references: nitrogen.csv/net_primary_production
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: milligram per kilogram per day
#>       valueURI: http://purl.obolibrary.org/obo/UO_0000311
#>   - references: nitrogen.csv/stem_mass_density
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: mass density
#>       valueURI: http://purl.dataone.org/odo/ECSO_00000678
#>   - references: nitrogen.csv/stem_mass_density
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: density unit
#>       valueURI: http://purl.obolibrary.org/obo/UO_0000182
#>   - references: nitrogen.csv/plant_density
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: Plant Material Biomass Density
#>       valueURI: http://purl.dataone.org/odo/ECSO_00001147
#>   - references: nitrogen.csv/plant_density
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: mass density unit
#>       valueURI: http://purl.obolibrary.org/obo/UO_0000052
#>   - references: nitrogen.csv/stem_n_concentration
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: nitrogen concentration
#>       valueURI: http://purl.dataone.org/odo/ECSO_00001883
#>   - references: nitrogen.csv/stem_n_concentration
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: concentration unit
#>       valueURI: http://purl.obolibrary.org/obo/UO_0000051
#>   - references: nitrogen.csv/n_assimilated
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: http://purl.dataone.org/odo/ECSO_00001883
#>       valueURI: http://purl.dataone.org/odo/ECSO_00001123
#>   - references: nitrogen.csv/n_assimilated
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: milligram per kilogram per day
#>       valueURI: http://purl.obolibrary.org/obo/UO_0000311
#>   - references: nitrogen.csv/site_name
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: station name
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002392
#>   - references: nitrogen.csv/site_lat
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: latitude coordinate
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002130
#>   - references: nitrogen.csv/site_lat
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: degree
#>       valueURI: http://purl.obolibrary.org/obo/UO_0000185
#>   - references: nitrogen.csv/site_lon
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: longitude coordinate
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002132
#>   - references: nitrogen.csv/site_lon
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: degree
#>       valueURI: http://purl.obolibrary.org/obo/UO_0000185
#>   - references: decomp.csv
#>     propertyURI:
#>       label: is about
#>       propertyURI: http://purl.obolibrary.org/obo/IAO_0000136
#>     valueURI:
#>       label: Leaf Litter Carbon Pool
#>       valueURI: http://purl.dataone.org/odo/ECSO_00000034
#>   - references: decomp.csv/type
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: type of soil core
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002623
#>   - references: decomp.csv/date
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: date and time of measurement
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002043
#>   - references: decomp.csv/date
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: year and day of measurement
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002072
#>   - references: decomp.csv/arm
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: site identifier
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002997
#>   - references: decomp.csv/ntrt
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: nitrate concentration
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002216
#>   - references: decomp.csv/ntrt
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: gram per square meter
#>       valueURI: http://purl.obolibrary.org/obo/UO_0010049
#>   - references: decomp.csv/year
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: year of measurement
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002050
#>   - references: decomp.csv/percent_loss
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: loss on ignition percentage
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002294
#>   - references: decomp.csv/percent_loss
#>     propertyURI:
#>       label: has unit
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>     valueURI:
#>       label: percent
#>       valueURI: http://purl.obolibrary.org/obo/UO_0000187
#>   - references: decomp.csv/taxa
#>     propertyURI:
#>       label: contains measurements of type
#>       propertyURI: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>     valueURI:
#>       label: taxonomic classification
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002730
#>   - references: ancillary_data.zip
#>     propertyURI:
#>       label: is about
#>       propertyURI: http://purl.obolibrary.org/obo/IAO_0000136
#>     valueURI:
#>       label: environmental feature
#>       valueURI: http://purl.obolibrary.org/obo/ENVO_00002297
#>   - references: processing_and_analysis.R
#>     propertyURI:
#>       label: is about
#>       propertyURI: http://purl.obolibrary.org/obo/IAO_0000136
#>     valueURI:
#>       label: analysis code
#>       valueURI: http://purl.dataone.org/odo/ECSO_00002489
#>   - references: Kelman R Wieder e9f6fad4-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Kelman R Wieder e9f6fad4-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Villanova University
#>       valueURI: https://ror.org/02g7kd627
#>   - references: Dale H Vitt e9f713f2-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Dale H Vitt e9f713f2-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Southern Illinois University Edwardsville
#>       valueURI: https://ror.org/04cqs5j56
#>   - references: Melanie A Vile e9f729b4-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Melanie A Vile e9f729b4-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Villanova University
#>       valueURI: https://ror.org/02g7kd627
#>   - references: Jeremy A Graham e9f7402a-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Jeremy A Graham e9f7402a-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Southern Illinois University Edwardsville
#>       valueURI: https://ror.org/04cqs5j56
#>   - references: Jeremy A Hartsock e9f75510-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Jeremy A Hartsock e9f75510-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Southern Illinois University Edwardsville
#>       valueURI: https://ror.org/04cqs5j56
#>   - references: Hope Fillingim e9f76ab4-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Hope Fillingim e9f76ab4-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Villanova University
#>       valueURI: https://ror.org/02g7kd627
#>   - references: Melissa House e9f78012-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Melissa House e9f78012-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Southern Illinois University Edwardsville
#>       valueURI: https://ror.org/04cqs5j56
#>   - references: James C Quinn e9f794a8-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: James C Quinn e9f794a8-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Villanova University
#>       valueURI: https://ror.org/02g7kd627
#>   - references: Kimberli D Scott e9f7aa4c-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Kimberli D Scott e9f7aa4c-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Villanova University
#>       valueURI: https://ror.org/02g7kd627
#>   - references: Meaghan Petix e9f7be9c-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Meaghan Petix e9f7be9c-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Southern Illinois University Edwardsville
#>       valueURI: https://ror.org/04cqs5j56
#>   - references: Kelly J McMillen e9f7d3f0-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Kelly J McMillen e9f7d3f0-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Villanova University
#>       valueURI: https://ror.org/02g7kd627
#>   - references: Kelman R Wieder e9f7e886-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Kelman R Wieder e9f7e886-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Villanova University
#>       valueURI: https://ror.org/02g7kd627
#>   - references: James C Quinn e9f80028-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: James C Quinn e9f80028-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: Villanova University
#>       valueURI: https://ror.org/02g7kd627
#>   - references: Jane Doe e9f81572-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: is a
#>       propertyURI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>     valueURI:
#>       label: Person
#>       valueURI: https://schema.org/Person
#>   - references: Jane Doe e9f81572-fa9f-11ef-acd1-0022484ed5a5
#>     propertyURI:
#>       label: member of
#>       propertyURI: https://schema.org/memberOf
#>     valueURI:
#>       label: University of Wisconsin–Madison
#>       valueURI: https://ror.org/01y2jtd41

# Remove the EML files from the temporary directory

unlink(paste0(tempdir(), "/edi.260.3.xml"))
unlink(paste0(tempdir(), "/edi.260.4.xml"))