Define categorical variables

Identify and define categorical variables for your data tables. Run this function whenever your data contain attributes of class "categorical" listed in attributes_datatablename.txt files.

define_catvars(path, data.path = path)

Arguments

path
(character) A path to your metadata templates directory containing attributes_datatablename.txt.

data.path
(character) A path to your data directory. Don't use this argument if your data are co-located with your metadata templates.

Value

A tab delimited UTF-8 file in the dataset working directory titled catvars_datatablename.txt containing unique codes of attributes of class "categorical". Open this file in a spreadsheet editor to supply definitions for listed codes.

Details

This function will not overwrite any catvars_datatablename.txt files you have created in the dataset working directory.

This function identifies "categorical" class attributes from the file attributes_datatablename.txt and extracts unique values of these attributes to the table for you to define.

Duplicate data file names are not allowed, even if they are different file types.