Import templates for geographic coverage

Extract detailed geographic coverage (latitude, longitude, and site name) to be included in the EML.

extract_geocoverage(path, data.path = path, data.file, lat.col,
    lon.col, site.col)

Arguments

path
A path to the metadata directory.

data.path
A path to the directory containing the data table with geographic information. Don't use this argument if the data table is located at the path argument listed above.

data.file
Name of the input data table containing geographic coverage data.

lat.col
Name of latitude column. Values of this column must be in decimal degrees. Latitudes south of the equator must be prefixed with a minus sign (i.e. dash, "-").

lon.col
Name of longitude column. Values of this column must be in decimal degrees. Longitudes west of the prime meridian must be prefixed with a minus sign (i.e. dash, "-").

site.col
Name of site column. This column lists site specific names to be associated with the geographic coordinates.

Value

A tab delimited file in the dataset working directory titled geographic_coverage.txt and containing decimal degree latitude, decimal degree longitude, and site name.