Render templates into EML metadata

Translate user supplied metadata templates into the EML, validate the EML, and write to file.

make_eml(
    path,
    data.path = path,
    eml.path = path,
    dataset.title,
    data.files,
    data.files.description,
    data.files.quote.character,
    data.files.url,
    zip.dir,
    zip.dir.description,
    temporal.coverage,
    geographic.description,
    geographic.coordinates,
    maintenance.description,
    user.id,
    affiliation,
    environment,
    package.id,
    provenance,
    return.obj = FALSE,
    write.file = TRUE)

Arguments

path
(character) A path to the directory containing the completed metadata templates.

data.path
(character) A path to the directory containing the data entities described by the metadata templates.

eml.path
(character) A path to the directory where the EML will be writtn.

dataset.title
(character) The title for your data package. Be descriptive (more than 5 words). We recommend the following format: Project name: Broad description: Time span (e.g. "GLEON: Long term lake chloride concentrations from North America and Europe: 1940-2016").

data.files
(character) A vector of character strings specifying the names of your data files (e.g. data.files = c("lake_chloride_concentrations.csv", "lake_characteristics.csv")).

data.files.description
(character) A vector of character strings briefly describing the data files listed in the data.files argument and in the same order as listed in the data.files argument (e.g. data.files.description = c("Chloride concentration data.", "Climate, road density, and impervious surface data."))

data.files.quote.character
(character) A vector of character strings defining the quote characters used in your data files and in the same order as listed in the data.files argument. This argument is required only if your data contain quotations. If the quote character is a quotation, then enter "\"". If the quote character is an apostrophe, then enter "\'". Example: data.files.quote.character = c("\"", "\"").

data.files.url
(character) The URL of where your data tables are stored on a publicly accessible server (i.e. does not require user ID or password). This argument is required only if your data are accessible from a publicly accesible URL. The EDI data repository software, PASTA+, will use this to upload your data into the repository. If you will be manually uploading your data tables, then don't use this argument. Example: data.files.url = "https://lter.limnology.wisc.edu/sites/default/files/data/gleon_chloride".

zip.dir
(character) A vector of character strings listing the .zip directories of your dataset.

zip.dir.description
(character) A vector of character strings briefly describing the contents of any .zip directory present in the working directory.

temporal.coverage
(character) A vector of character strings specifying the beginning and ending dates of your dataset. Use the format YYYY-MM-DD.

geographic.description
(character) A description of the geographic coverage of your dataset. Don't use this argument if you are supplying geographic.coordinates in the bounding_boxes.txt template file. Example: "North America and Europe".

geographic.coordinates
(character) A vector of character strings specifying the spatial bounding coordinates of your dataset in decimal degrees. This argument is not required if you are supplying bounding coordinates in the bounding_boxes.txt template file. The list must follow this order: North, East, South, West. Longitudes West of the prime meridian and latitudes South of the equator are prefixed with a minus sign (i.e. dash -). If you don't have an area, but rather a point, Repeat the latitude value for North and South, and repeat the longitude value for East and West (e.g. geographic.coordinates = c('28.38', '-119.95', '28.38', '-119.95)).

maintenance.description
(character) A description of whether data collection for this dataset is "ongoing" or "completed".

user.id
(character) A vector of character strings, specifying your user ID for the EDI data repository. The user.id controls editing access to your data package. If you do not have one, contact EDI (info@environmentaldatainitiative.org) to obtain one. In the meantime do not use this argument when running `make_eml`.

affiliation
(character) A vector of character strings, specifying the affiliation of your user ID. In a list, the associations must follow the same order of the corresponding values listed under user.id. This is the affiliation used when logging in to the EDI Data Portal and can be: "LTER" or "EDI". If you don't have a user.id then do not use this argument when running `make_eml`.

environment
(character) Data repository environment to create the package in. Can be: 'development', 'staging', 'production'.

package.id
(character) The ID of your data package. A missing package ID defaults to edi.101.1. A package ID must contain the scope, package number, and revision number (e.g. 'edi.101.1').

provenance
(character) A vector of EDI package.ids corresponding to parent data packages from which this data package was created. A package ID must contain the scope, package number, and revision number (e.g. 'knb-lter-cap.46.3').

return.obj
(logical) Return the EML as an EML object for further use in the R environment.

write.file
(logical)

Value

Validation results printed to the RStudio Console.

An EML file written to the dataset working directory titled packageID.xml.

Details

If validation fails, open the EML document in a .xml editor to identify the source of error. Often the error is small and quickly resolved with the aid of an editors schema validator.