Read templates and data files into a list structure as an alternative input to `EMLassemblyline` functions

Use this function to read metadata templates and data into an R list object for input to `EMLassemblyline` functions, rather than supplying the files themselves.

This function is primarily used for testing and demonstration purposes however, this approach is generally useful for inputing metadata and data from a meta- or data-base and other upstream sources. See below for list object structure.

read_files(path, data.path = path, data.table = NULL,
other.entity = NULL, sep = NULL)

Arguments

path
(character) Path to metadata templates.

data.path
(character) Path to data tables and other entities.

data.table
(character) Data table name(s). If more than one, then supply as a vector of character strings (e.g. `data.table = c("concentrations.csv", "characteristics.csv")`).

other.entity
(character) Name(s) of `other.entity`(s) in this dataset. Use `other.entity` for all non-`data.table` files. `other.entity`(s) should be stored at `data.path`.

sep
(character) Delimiter of `data.table`. Use this argument if `read_files` fails to automatically identify `data.table` field delimiters.

Value

(named list) Named list with this internal structure (NOTE: use `NA` for non-existant content):

`template` Metadata template files

`abstract.txt` Abstract template

`content` EML::TextType containing abstract

`path` Path to file

`additional_info.txt` Additional information template

`content` EML::TextType containing additional info

`path` Path to file

`attributes_*.txt` Attributes template where * is name of `data.table` described by the attributes

`content` Data frame of attributes template

`path` Path to file

`bounding_boxes.txt` Bounding boxes template

`content` Data frame of bounding boxes template

`path` Path to file

`catvars_*.txt` Categorical variables template where * is name of `data.table` described by the catvars

`content` Data frame of categorical variables template

`path` Path to file

`custom_units.txt` Custom units template

`content` Data frame of custom units template

`path` Path to file

`geographic_coverage.txt` Geographic coverage template

`content` Data frame of geographic coverage template

`path` Path to file

`intellectual_rights.txt` Intellectual rights template

`content` EML::TextType containing intellectual rights

`path` Path to file

`keywords.txt` Keywords template

`content` Data frame of keywords template

`path` Path to file

`methods.txt` Methods template

`content` EML::Methods containing methods

`path` Path to file

`personnel.txt` Personnel template

`content` Data frame of personnel template

`path` Path to file

`taxonomicCoverage.xml` Taxonomic coverage EML element

`content` EML::taxonomicCoverage

`path` Path to file

`data.table`

`data.table.name.ext` Name of data table including file extension

`content` Data frame of data table

`path` Path to data table (required)

`other.entity`

`other.entity.name.ext` Name of other entity including file extension

`content` Set to `NA`

`path` Path to other entity (required)

Note

Paths to data tables and other entities must be present for auto-generated metadata content.