Create the annotations template

The annotations template (annotations.txt) lists the elements and corresponding ontology terms for annotating your EML. Run this function after all your EMLassemblyline metadata templates are complete or if you're annotating an existing EML file. This function gathers annotatable elements from your EML and assigns default predicate labels and URIs. You must provide object labels and URIs from the ontology of your choosing.

template_annotations(
  path,
  data.path = path,
  data.table = NULL,
  other.entity = NULL,
  default.annotations = NULL,
  eml.path = path,
  eml = NULL
)

Arguments

path
(character) Path to the metadata template directory and where annotations.txt will be written.

data.path
(character; optional) Path to the data directory. Defaults to path.

data.table
(character; optional) Table name. If more than one, then supply as a vector of character strings (e.g. data.table = c('nitrogen.csv', 'decomp.csv')).

other.entity
(character; optional) Other entity name. If more than one, then supply as a vector of character strings (e.g. other.entity = c('maps.zip', 'analysis.R')).

default.annotations
(data frame; optional) Default annotations added to annotations.txt. EMLassemblyline specified defaults are used unless specifying your own (see note below). You can manually change these after annotations.txt has been created.

eml.path
(character; optional) Path to the EML directory. Use this if creating annotations.txt for an EML file. Defaults to path.

eml
(EML file; optional) An EML file located at eml.path. Use this if annotating an existing EML file.

Value

annotations.txt The tab delimited annotations template, pre-populated with default annotations, and containing the fields:

id - A unique identifier for the element being annotated.

element - The element being annotated.

context - The context of the subject (i.e. element value) being annotated (e.g. If the same column name occurs in more than one data tables, you will need to know which table it came from.).

subject - The element value to be annotated.

predicate_label - The predicate label (a.k.a. property) describing the relation of the subject to the object. This label should be copied directly from an ontology.

predicate_uri - The predicate label URI copied directly from an ontology.

object_label - The object label (a.k.a. value) describing the subject. This label should be copied directly from an ontology.

object_uri - The object URI copied from an ontology.

The general user should ignore the id and element fields and focus on the subject, context predicate_label, predicate_uri, object_label, and object_uri fields. Only the predicate and object fields should be modified. If you want to add an annotation to any of the listed subjects, simply copy the full row containing the subject, paste it in as new line, and modify the predicate and object fields.

Note

To set your own default annotations, copy the EMLassemblyline defaults (file.copy(from = system.file("/templates/annotation_defaults.txt", package = "EMLassemblyline"), to = path), where path is where you want the file written to) then change the values in the predicate_label, predicate_uri, object_label, and object_uri fields, save the file, read it in to R as a data frame and use it with the default.annotations argument.

Some users may want to build annotations.txt from scratch. A few rules to follow when doing this:

id - IDs must be unique for each unique subject.

element - Supported elements and the required syntax is listed under the element column of annotation_defaults.txt. View this file with View(system.file("/templates/annotation_defaults.txt", package = "EMLassemblyline"))

context - Context values are only required for elements that are nested within other elements. Currently only /dataTable/attribute elements require context where the dataTable objectName is the context (e.g. nitrogen.csv).

subject - Subjects are required for each annotation. For /dataset the subject is "dataset". For /dataTable the subject is the file name. For /dataTable/attribute the subjects are the dataTable field names. For /otherEntity the subject is the file name. For /ResponsibleParty the subject is created with paste(first.name, middle.name, last.name, collapse = " ")

Examples

# From a set of EMLassemblyline metadata templates ----------------------------
#
# Create annotations.txt for a dataset of data tables, other entities, and 
# described by EMLassemblyline metadata templates

# Copy templates and data objects to a temporary directory

file.copy(
  from = system.file("/examples/pkg_260", package = "EMLassemblyline"),
  to = tempdir(),
  recursive = TRUE
)

#> [1] TRUE

unlink(paste0(tempdir(), "/pkg_260/metadata_templates/annotations.txt"))

# Create annotations.txt

template_annotations(
 path = paste0(tempdir(), "/pkg_260/metadata_templates"),
 data.path = paste0(tempdir(), "/pkg_260/data_objects"),
 data.table = c("nitrogen.csv", "decomp.csv"),
 other.entity = c("ancillary_data.zip", "processing_and_analysis.R")
)

#> Templating annotations ...

#> Done.

# View contents

df <- data.table::fread(
  paste0(tempdir(), "/pkg_260/metadata_templates/annotations.txt")
)
df

#>                                       id              element      context
#>  1:                             /dataset             /dataset          eml
#>  2:                        /nitrogen.csv           /dataTable      dataset
#>  3:                   /nitrogen.csv/date /dataTable/attribute nitrogen.csv
#>  4:                   /nitrogen.csv/date /dataTable/attribute nitrogen.csv
#>  5:                   /nitrogen.csv/ntrt /dataTable/attribute nitrogen.csv
#>  6:                   /nitrogen.csv/ntrt /dataTable/attribute nitrogen.csv
#>  7:                    /nitrogen.csv/arm /dataTable/attribute nitrogen.csv
#>  8:                    /nitrogen.csv/arm /dataTable/attribute nitrogen.csv
#>  9: /nitrogen.csv/net_primary_production /dataTable/attribute nitrogen.csv
#> 10: /nitrogen.csv/net_primary_production /dataTable/attribute nitrogen.csv
#> 11:      /nitrogen.csv/stem_mass_density /dataTable/attribute nitrogen.csv
#> 12:      /nitrogen.csv/stem_mass_density /dataTable/attribute nitrogen.csv
#> 13:          /nitrogen.csv/plant_density /dataTable/attribute nitrogen.csv
#> 14:          /nitrogen.csv/plant_density /dataTable/attribute nitrogen.csv
#> 15:   /nitrogen.csv/stem_n_concentration /dataTable/attribute nitrogen.csv
#> 16:   /nitrogen.csv/stem_n_concentration /dataTable/attribute nitrogen.csv
#> 17:          /nitrogen.csv/n_assimilated /dataTable/attribute nitrogen.csv
#> 18:          /nitrogen.csv/n_assimilated /dataTable/attribute nitrogen.csv
#> 19:              /nitrogen.csv/site_name /dataTable/attribute nitrogen.csv
#> 20:              /nitrogen.csv/site_name /dataTable/attribute nitrogen.csv
#> 21:               /nitrogen.csv/site_lat /dataTable/attribute nitrogen.csv
#> 22:               /nitrogen.csv/site_lat /dataTable/attribute nitrogen.csv
#> 23:               /nitrogen.csv/site_lon /dataTable/attribute nitrogen.csv
#> 24:               /nitrogen.csv/site_lon /dataTable/attribute nitrogen.csv
#> 25:                          /decomp.csv           /dataTable      dataset
#> 26:                     /decomp.csv/type /dataTable/attribute   decomp.csv
#> 27:                     /decomp.csv/type /dataTable/attribute   decomp.csv
#> 28:                     /decomp.csv/date /dataTable/attribute   decomp.csv
#> 29:                     /decomp.csv/date /dataTable/attribute   decomp.csv
#> 30:                      /decomp.csv/arm /dataTable/attribute   decomp.csv
#> 31:                      /decomp.csv/arm /dataTable/attribute   decomp.csv
#> 32:                     /decomp.csv/ntrt /dataTable/attribute   decomp.csv
#> 33:                     /decomp.csv/ntrt /dataTable/attribute   decomp.csv
#> 34:                     /decomp.csv/year /dataTable/attribute   decomp.csv
#> 35:                     /decomp.csv/year /dataTable/attribute   decomp.csv
#> 36:             /decomp.csv/percent_loss /dataTable/attribute   decomp.csv
#> 37:             /decomp.csv/percent_loss /dataTable/attribute   decomp.csv
#> 38:                     /decomp.csv/taxa /dataTable/attribute   decomp.csv
#> 39:                     /decomp.csv/taxa /dataTable/attribute   decomp.csv
#> 40:                  /ancillary_data.zip         /otherEntity      dataset
#> 41:           /processing_and_analysis.R         /otherEntity      dataset
#> 42:                     /Kelman R Wieder    /ResponsibleParty      dataset
#> 43:                     /Kelman R Wieder    /ResponsibleParty      dataset
#> 44:                         /Dale H Vitt    /ResponsibleParty      dataset
#> 45:                         /Dale H Vitt    /ResponsibleParty      dataset
#> 46:                      /Melanie A Vile    /ResponsibleParty      dataset
#> 47:                      /Melanie A Vile    /ResponsibleParty      dataset
#> 48:                     /Jeremy A Graham    /ResponsibleParty      dataset
#> 49:                     /Jeremy A Graham    /ResponsibleParty      dataset
#> 50:                   /Jeremy A Hartsock    /ResponsibleParty      dataset
#> 51:                   /Jeremy A Hartsock    /ResponsibleParty      dataset
#> 52:                      /Hope Fillingim    /ResponsibleParty      dataset
#> 53:                      /Hope Fillingim    /ResponsibleParty      dataset
#> 54:                       /Melissa House    /ResponsibleParty      dataset
#> 55:                       /Melissa House    /ResponsibleParty      dataset
#> 56:                       /James C Quinn    /ResponsibleParty      dataset
#> 57:                       /James C Quinn    /ResponsibleParty      dataset
#> 58:                    /Kimberli D Scott    /ResponsibleParty      dataset
#> 59:                    /Kimberli D Scott    /ResponsibleParty      dataset
#> 60:                       /Meaghan Petix    /ResponsibleParty      dataset
#> 61:                       /Meaghan Petix    /ResponsibleParty      dataset
#> 62:                    /Kelly J McMillen    /ResponsibleParty      dataset
#> 63:                    /Kelly J McMillen    /ResponsibleParty      dataset
#> 64:                            /Jane Doe    /ResponsibleParty      dataset
#> 65:                            /Jane Doe    /ResponsibleParty      dataset
#> 66:                           /Joe Smith    /ResponsibleParty      dataset
#> 67:                           /Joe Smith    /ResponsibleParty      dataset
#> 68:                         /Chase Gacho    /ResponsibleParty      dataset
#> 69:                         /Chase Gacho    /ResponsibleParty      dataset
#>                                       id              element      context
#>                       subject               predicate_label
#>  1:                   dataset                      is about
#>  2:              nitrogen.csv                      is about
#>  3:                      date contains measurements of type
#>  4:                      date                      has unit
#>  5:                      ntrt contains measurements of type
#>  6:                      ntrt                      has unit
#>  7:                       arm contains measurements of type
#>  8:                       arm                      has unit
#>  9:    net_primary_production contains measurements of type
#> 10:    net_primary_production                      has unit
#> 11:         stem_mass_density contains measurements of type
#> 12:         stem_mass_density                      has unit
#> 13:             plant_density contains measurements of type
#> 14:             plant_density                      has unit
#> 15:      stem_n_concentration contains measurements of type
#> 16:      stem_n_concentration                      has unit
#> 17:             n_assimilated contains measurements of type
#> 18:             n_assimilated                      has unit
#> 19:                 site_name contains measurements of type
#> 20:                 site_name                      has unit
#> 21:                  site_lat contains measurements of type
#> 22:                  site_lat                      has unit
#> 23:                  site_lon contains measurements of type
#> 24:                  site_lon                      has unit
#> 25:                decomp.csv                      is about
#> 26:                      type contains measurements of type
#> 27:                      type                      has unit
#> 28:                      date contains measurements of type
#> 29:                      date                      has unit
#> 30:                       arm contains measurements of type
#> 31:                       arm                      has unit
#> 32:                      ntrt contains measurements of type
#> 33:                      ntrt                      has unit
#> 34:                      year contains measurements of type
#> 35:                      year                      has unit
#> 36:              percent_loss contains measurements of type
#> 37:              percent_loss                      has unit
#> 38:                      taxa contains measurements of type
#> 39:                      taxa                      has unit
#> 40:        ancillary_data.zip                      is about
#> 41: processing_and_analysis.R                      is about
#> 42:           Kelman R Wieder                          is a
#> 43:           Kelman R Wieder                     member of
#> 44:               Dale H Vitt                          is a
#> 45:               Dale H Vitt                     member of
#> 46:            Melanie A Vile                          is a
#> 47:            Melanie A Vile                     member of
#> 48:           Jeremy A Graham                          is a
#> 49:           Jeremy A Graham                     member of
#> 50:         Jeremy A Hartsock                          is a
#> 51:         Jeremy A Hartsock                     member of
#> 52:            Hope Fillingim                          is a
#> 53:            Hope Fillingim                     member of
#> 54:             Melissa House                          is a
#> 55:             Melissa House                     member of
#> 56:             James C Quinn                          is a
#> 57:             James C Quinn                     member of
#> 58:          Kimberli D Scott                          is a
#> 59:          Kimberli D Scott                     member of
#> 60:             Meaghan Petix                          is a
#> 61:             Meaghan Petix                     member of
#> 62:          Kelly J McMillen                          is a
#> 63:          Kelly J McMillen                     member of
#> 64:                  Jane Doe                          is a
#> 65:                  Jane Doe                     member of
#> 66:                 Joe Smith                          is a
#> 67:                 Joe Smith                     member of
#> 68:               Chase Gacho                          is a
#> 69:               Chase Gacho                     member of
#>                       subject               predicate_label
#>                                                                        predicate_uri
#>  1:                                       http://purl.obolibrary.org/obo/IAO_0000136
#>  2:                                       http://purl.obolibrary.org/obo/IAO_0000136
#>  3: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>  4:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>  5: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>  6:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>  7: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>  8:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>  9: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 10:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 11: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 12:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 13: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 14:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 15: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 16:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 17: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 18:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 19: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 20:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 21: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 22:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 23: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 24:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 25:                                       http://purl.obolibrary.org/obo/IAO_0000136
#> 26: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 27:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 28: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 29:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 30: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 31:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 32: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 33:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 34: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 35:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 36: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 37:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 38: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 39:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 40:                                       http://purl.obolibrary.org/obo/IAO_0000136
#> 41:                                       http://purl.obolibrary.org/obo/IAO_0000136
#> 42:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 43:                                                      https://schema.org/memberOf
#> 44:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 45:                                                      https://schema.org/memberOf
#> 46:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 47:                                                      https://schema.org/memberOf
#> 48:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 49:                                                      https://schema.org/memberOf
#> 50:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 51:                                                      https://schema.org/memberOf
#> 52:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 53:                                                      https://schema.org/memberOf
#> 54:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 55:                                                      https://schema.org/memberOf
#> 56:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 57:                                                      https://schema.org/memberOf
#> 58:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 59:                                                      https://schema.org/memberOf
#> 60:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 61:                                                      https://schema.org/memberOf
#> 62:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 63:                                                      https://schema.org/memberOf
#> 64:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 65:                                                      https://schema.org/memberOf
#> 66:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 67:                                                      https://schema.org/memberOf
#> 68:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 69:                                                      https://schema.org/memberOf
#>                                                                        predicate_uri
#>     object_label                object_uri
#>  1:                                       
#>  2:                                       
#>  3:                                       
#>  4:                                       
#>  5:                                       
#>  6:                                       
#>  7:                                       
#>  8:                                       
#>  9:                                       
#> 10:                                       
#> 11:                                       
#> 12:                                       
#> 13:                                       
#> 14:                                       
#> 15:                                       
#> 16:                                       
#> 17:                                       
#> 18:                                       
#> 19:                                       
#> 20:                                       
#> 21:                                       
#> 22:                                       
#> 23:                                       
#> 24:                                       
#> 25:                                       
#> 26:                                       
#> 27:                                       
#> 28:                                       
#> 29:                                       
#> 30:                                       
#> 31:                                       
#> 32:                                       
#> 33:                                       
#> 34:                                       
#> 35:                                       
#> 36:                                       
#> 37:                                       
#> 38:                                       
#> 39:                                       
#> 40:                                       
#> 41:                                       
#> 42:       Person https://schema.org/Person
#> 43:                                       
#> 44:       Person https://schema.org/Person
#> 45:                                       
#> 46:       Person https://schema.org/Person
#> 47:                                       
#> 48:       Person https://schema.org/Person
#> 49:                                       
#> 50:       Person https://schema.org/Person
#> 51:                                       
#> 52:       Person https://schema.org/Person
#> 53:                                       
#> 54:       Person https://schema.org/Person
#> 55:                                       
#> 56:       Person https://schema.org/Person
#> 57:                                       
#> 58:       Person https://schema.org/Person
#> 59:                                       
#> 60:       Person https://schema.org/Person
#> 61:                                       
#> 62:       Person https://schema.org/Person
#> 63:                                       
#> 64:       Person https://schema.org/Person
#> 65:                                       
#> 66:       Person https://schema.org/Person
#> 67:                                       
#> 68:       Person https://schema.org/Person
#> 69:                                       
#>     object_label                object_uri

# Clean up temporary directory

unlink(paste0(tempdir(), "/pkg_260"), recursive = TRUE, force = TRUE)

# From an EML file ------------------------------------------------------------
#
# Create annotations.txt for an EML file

# Copy an EML file to a temporary directory

file.copy(
  from = system.file("/examples/eml/edi.260.3.xml", package = "EMLassemblyline"),
  to = tempdir(),
  recursive = TRUE
)

#> [1] TRUE

# Create annotations.txt

template_annotations(
 path = tempdir(),
 eml = "edi.260.3.xml"
)

#> Templating annotations ...

#> Done.

# View contents

df <- data.table::fread(
 paste0(tempdir(), "/annotations.txt")
)
df

#>                                       id              element      context
#>  1:                             /dataset             /dataset          eml
#>  2:                          /decomp.csv           /dataTable      dataset
#>  3:                     /decomp.csv/type /dataTable/attribute   decomp.csv
#>  4:                     /decomp.csv/type /dataTable/attribute   decomp.csv
#>  5:                     /decomp.csv/date /dataTable/attribute   decomp.csv
#>  6:                     /decomp.csv/date /dataTable/attribute   decomp.csv
#>  7:                      /decomp.csv/arm /dataTable/attribute   decomp.csv
#>  8:                      /decomp.csv/arm /dataTable/attribute   decomp.csv
#>  9:                     /decomp.csv/ntrt /dataTable/attribute   decomp.csv
#> 10:                     /decomp.csv/ntrt /dataTable/attribute   decomp.csv
#> 11:                     /decomp.csv/year /dataTable/attribute   decomp.csv
#> 12:                     /decomp.csv/year /dataTable/attribute   decomp.csv
#> 13:             /decomp.csv/percent_loss /dataTable/attribute   decomp.csv
#> 14:             /decomp.csv/percent_loss /dataTable/attribute   decomp.csv
#> 15:                     /decomp.csv/taxa /dataTable/attribute   decomp.csv
#> 16:                     /decomp.csv/taxa /dataTable/attribute   decomp.csv
#> 17:                        /nitrogen.csv           /dataTable      dataset
#> 18:                   /nitrogen.csv/date /dataTable/attribute nitrogen.csv
#> 19:                   /nitrogen.csv/date /dataTable/attribute nitrogen.csv
#> 20:                   /nitrogen.csv/ntrt /dataTable/attribute nitrogen.csv
#> 21:                   /nitrogen.csv/ntrt /dataTable/attribute nitrogen.csv
#> 22:                    /nitrogen.csv/arm /dataTable/attribute nitrogen.csv
#> 23:                    /nitrogen.csv/arm /dataTable/attribute nitrogen.csv
#> 24: /nitrogen.csv/net_primary_production /dataTable/attribute nitrogen.csv
#> 25: /nitrogen.csv/net_primary_production /dataTable/attribute nitrogen.csv
#> 26:      /nitrogen.csv/stem_mass_density /dataTable/attribute nitrogen.csv
#> 27:      /nitrogen.csv/stem_mass_density /dataTable/attribute nitrogen.csv
#> 28:          /nitrogen.csv/plant_density /dataTable/attribute nitrogen.csv
#> 29:          /nitrogen.csv/plant_density /dataTable/attribute nitrogen.csv
#> 30:   /nitrogen.csv/stem_n_concentration /dataTable/attribute nitrogen.csv
#> 31:   /nitrogen.csv/stem_n_concentration /dataTable/attribute nitrogen.csv
#> 32:          /nitrogen.csv/n_assimilated /dataTable/attribute nitrogen.csv
#> 33:          /nitrogen.csv/n_assimilated /dataTable/attribute nitrogen.csv
#> 34:              /nitrogen.csv/site_name /dataTable/attribute nitrogen.csv
#> 35:              /nitrogen.csv/site_name /dataTable/attribute nitrogen.csv
#> 36:               /nitrogen.csv/site_lat /dataTable/attribute nitrogen.csv
#> 37:               /nitrogen.csv/site_lat /dataTable/attribute nitrogen.csv
#> 38:               /nitrogen.csv/site_lon /dataTable/attribute nitrogen.csv
#> 39:               /nitrogen.csv/site_lon /dataTable/attribute nitrogen.csv
#> 40:                  /ancillary_data.zip         /otherEntity      dataset
#> 41:           /processing_and_analysis.R         /otherEntity      dataset
#> 42:                     /Wieder Kelman R    /ResponsibleParty      dataset
#> 43:                     /Wieder Kelman R    /ResponsibleParty      dataset
#> 44:                         /Vitt Dale H    /ResponsibleParty      dataset
#> 45:                         /Vitt Dale H    /ResponsibleParty      dataset
#> 46:                      /Vile Melanie A    /ResponsibleParty      dataset
#> 47:                      /Vile Melanie A    /ResponsibleParty      dataset
#> 48:                     /Graham Jeremy A    /ResponsibleParty      dataset
#> 49:                     /Graham Jeremy A    /ResponsibleParty      dataset
#> 50:                   /Hartsock Jeremy A    /ResponsibleParty      dataset
#> 51:                   /Hartsock Jeremy A    /ResponsibleParty      dataset
#> 52:                      /Hope Fillingim    /ResponsibleParty      dataset
#> 53:                      /Hope Fillingim    /ResponsibleParty      dataset
#> 54:                       /Melissa House    /ResponsibleParty      dataset
#> 55:                       /Melissa House    /ResponsibleParty      dataset
#> 56:                       /Quinn James C    /ResponsibleParty      dataset
#> 57:                       /Quinn James C    /ResponsibleParty      dataset
#> 58:                    /Scott Kimberli D    /ResponsibleParty      dataset
#> 59:                    /Scott Kimberli D    /ResponsibleParty      dataset
#> 60:                       /Meaghan Petix    /ResponsibleParty      dataset
#> 61:                       /Meaghan Petix    /ResponsibleParty      dataset
#> 62:                    /McMillen Kelly J    /ResponsibleParty      dataset
#> 63:                    /McMillen Kelly J    /ResponsibleParty      dataset
#> 64:                            /Jane Doe    /ResponsibleParty      dataset
#> 65:                            /Jane Doe    /ResponsibleParty      dataset
#>                                       id              element      context
#>                       subject               predicate_label
#>  1:                   dataset                      is about
#>  2:                decomp.csv                      is about
#>  3:                      type contains measurements of type
#>  4:                      type                      has unit
#>  5:                      date contains measurements of type
#>  6:                      date                      has unit
#>  7:                       arm contains measurements of type
#>  8:                       arm                      has unit
#>  9:                      ntrt contains measurements of type
#> 10:                      ntrt                      has unit
#> 11:                      year contains measurements of type
#> 12:                      year                      has unit
#> 13:              percent_loss contains measurements of type
#> 14:              percent_loss                      has unit
#> 15:                      taxa contains measurements of type
#> 16:                      taxa                      has unit
#> 17:              nitrogen.csv                      is about
#> 18:                      date contains measurements of type
#> 19:                      date                      has unit
#> 20:                      ntrt contains measurements of type
#> 21:                      ntrt                      has unit
#> 22:                       arm contains measurements of type
#> 23:                       arm                      has unit
#> 24:    net_primary_production contains measurements of type
#> 25:    net_primary_production                      has unit
#> 26:         stem_mass_density contains measurements of type
#> 27:         stem_mass_density                      has unit
#> 28:             plant_density contains measurements of type
#> 29:             plant_density                      has unit
#> 30:      stem_n_concentration contains measurements of type
#> 31:      stem_n_concentration                      has unit
#> 32:             n_assimilated contains measurements of type
#> 33:             n_assimilated                      has unit
#> 34:                 site_name contains measurements of type
#> 35:                 site_name                      has unit
#> 36:                  site_lat contains measurements of type
#> 37:                  site_lat                      has unit
#> 38:                  site_lon contains measurements of type
#> 39:                  site_lon                      has unit
#> 40:        ancillary_data.zip                      is about
#> 41: processing_and_analysis.R                      is about
#> 42:           Wieder Kelman R                          is a
#> 43:           Wieder Kelman R                     member of
#> 44:               Vitt Dale H                          is a
#> 45:               Vitt Dale H                     member of
#> 46:            Vile Melanie A                          is a
#> 47:            Vile Melanie A                     member of
#> 48:           Graham Jeremy A                          is a
#> 49:           Graham Jeremy A                     member of
#> 50:         Hartsock Jeremy A                          is a
#> 51:         Hartsock Jeremy A                     member of
#> 52:            Hope Fillingim                          is a
#> 53:            Hope Fillingim                     member of
#> 54:             Melissa House                          is a
#> 55:             Melissa House                     member of
#> 56:             Quinn James C                          is a
#> 57:             Quinn James C                     member of
#> 58:          Scott Kimberli D                          is a
#> 59:          Scott Kimberli D                     member of
#> 60:             Meaghan Petix                          is a
#> 61:             Meaghan Petix                     member of
#> 62:          McMillen Kelly J                          is a
#> 63:          McMillen Kelly J                     member of
#> 64:                  Jane Doe                          is a
#> 65:                  Jane Doe                     member of
#>                       subject               predicate_label
#>                                                                        predicate_uri
#>  1:                                       http://purl.obolibrary.org/obo/IAO_0000136
#>  2:                                       http://purl.obolibrary.org/obo/IAO_0000136
#>  3: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>  4:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>  5: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>  6:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>  7: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#>  8:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#>  9: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 10:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 11: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 12:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 13: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 14:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 15: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 16:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 17:                                       http://purl.obolibrary.org/obo/IAO_0000136
#> 18: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 19:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 20: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 21:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 22: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 23:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 24: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 25:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 26: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 27:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 28: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 29:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 30: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 31:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 32: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 33:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 34: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 35:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 36: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 37:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 38: http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#containsMeasurementsOfType
#> 39:                    http://ecoinformatics.org/oboe/oboe.1.2/oboe-core.owl#hasUnit
#> 40:                                       http://purl.obolibrary.org/obo/IAO_0000136
#> 41:                                       http://purl.obolibrary.org/obo/IAO_0000136
#> 42:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 43:                                                      https://schema.org/memberOf
#> 44:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 45:                                                      https://schema.org/memberOf
#> 46:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 47:                                                      https://schema.org/memberOf
#> 48:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 49:                                                      https://schema.org/memberOf
#> 50:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 51:                                                      https://schema.org/memberOf
#> 52:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 53:                                                      https://schema.org/memberOf
#> 54:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 55:                                                      https://schema.org/memberOf
#> 56:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 57:                                                      https://schema.org/memberOf
#> 58:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 59:                                                      https://schema.org/memberOf
#> 60:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 61:                                                      https://schema.org/memberOf
#> 62:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 63:                                                      https://schema.org/memberOf
#> 64:                                  http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 65:                                                      https://schema.org/memberOf
#>                                                                        predicate_uri
#>     object_label                object_uri
#>  1:                                       
#>  2:                                       
#>  3:                                       
#>  4:                                       
#>  5:                                       
#>  6:                                       
#>  7:                                       
#>  8:                                       
#>  9:                                       
#> 10:                                       
#> 11:                                       
#> 12:                                       
#> 13:                                       
#> 14:                                       
#> 15:                                       
#> 16:                                       
#> 17:                                       
#> 18:                                       
#> 19:                                       
#> 20:                                       
#> 21:                                       
#> 22:                                       
#> 23:                                       
#> 24:                                       
#> 25:                                       
#> 26:                                       
#> 27:                                       
#> 28:                                       
#> 29:                                       
#> 30:                                       
#> 31:                                       
#> 32:                                       
#> 33:                                       
#> 34:                                       
#> 35:                                       
#> 36:                                       
#> 37:                                       
#> 38:                                       
#> 39:                                       
#> 40:                                       
#> 41:                                       
#> 42:       Person https://schema.org/Person
#> 43:                                       
#> 44:       Person https://schema.org/Person
#> 45:                                       
#> 46:       Person https://schema.org/Person
#> 47:                                       
#> 48:       Person https://schema.org/Person
#> 49:                                       
#> 50:       Person https://schema.org/Person
#> 51:                                       
#> 52:       Person https://schema.org/Person
#> 53:                                       
#> 54:       Person https://schema.org/Person
#> 55:                                       
#> 56:       Person https://schema.org/Person
#> 57:                                       
#> 58:       Person https://schema.org/Person
#> 59:                                       
#> 60:       Person https://schema.org/Person
#> 61:                                       
#> 62:       Person https://schema.org/Person
#> 63:                                       
#> 64:       Person https://schema.org/Person
#> 65:                                       
#>     object_label                object_uri

# Clean up temporary directory

unlink(paste0(tempdir(), "/edi.260.3.xml"), force = TRUE)
unlink(paste0(tempdir(), "/annotations.txt"), force = TRUE)