Create taxonomic coverage template

Extract unique taxa and resolve to one or more taxonomic authorities. make_eml will use this information to get the taxonomic hierarchy for each taxa and add to the EML. The metadata resulting from this function allows searching on any rank of the taxa.

template_taxonomic_coverage(
      path,
      data.path = path,
      taxa.table,
      taxa.col,
      taxa.name.type,
      taxa.authority,
      x = NULL,
      write.file = TRUE
    )

Arguments

path
(character) Path to where the template will be written.

data.path
(character) Path to where taxa.table is stored.

taxa.table
(character) Name of data table containing taxa.col.

taxa.col
(character) Name of column in taxa.table containing taxa names. Names can be single words or species binomials.

taxa.name.type
(character) Type of taxa names in taxa.col Can be: scientific, common, or both.

taxa.authority
(integer) An ordered numeric vector of ID's corresponding to data sources (i.e. taxonomic authorities) you'd like to resolve taxa names to, in the order of decreasing preference. See the list of supported data sources with view_taxa_authorities. Columns "resolve_sci_taxa", and "resolve_comm_taxa" list authorites supporting scientific and common searches, respectively.

x
(named list) Alternative input/output to EMLassemblyline functions. Use template_arguments to create x.

write.file
(logical) Write "taxonomic_coverage.txt" to path.

Value

taxonomic_coverage.txt A tab delimited file written to path containing authority system names and authority IDs for successfully resolved taxa, and "NA" otherwise.

If using x, then content of "taxonomic_coverage.txt is added to x under "/x/templates".

Details

template_taxonomic_coverage searches the most preferred taxonomic authority for all unique taxa listed in taxa.col returning the authority name and corresponding taxa identifier for direct matches (no fuzzy searching), then the next most preferred taxonomic authority is search for taxa that have not yet been resolved. This process repeats for subsequently listed authorities. "NA" is returned when an authority match is not made.

When "taxonomic_coverage.txt" is passed to make_eml, the authority information is used to get the hierarchical rank names of resolved taxa and rendered into the "taxonomicCoverage" element of EML.

Existing "taxonomic_coverage.txt" will not be overwritten by subsequent calls to template_taxonomic_coverage.