Import core metadata and table attribute templates

Import template files for storage of core metadata and table attributes. Some templates are populated with content detected by automated metadata extraction methods. The remainder is manually entered by the user. Instructions for editing the templates.

import_templates(
      path,
      data.path = path,
      license,
      data.table = NULL,
      write.file = TRUE,
      x = NULL
    )

Arguments

path
(character) Path to the directory where the templates will be written.

data.path
(character) Path to the directory where the data tables are stored.

license
(character) License under which the data will be released. Use https://creativecommons.org/publicdomain/zero/1.0/ or https://creativecommons.org/licenses/by/4.0/.

data.table
(character) Name of data table. If more than one, then supply as a vector of character strings (e.g. data.table = c('concentrations.csv', 'characteristics.csv')).

write.file
(logical; optional) Whether to write the templates to path.

x
(named list; optional) Alternative input to EMLassemblyline functions. Use template_arguments() to create x.

Value

abstract.txt Template for the dataset abstract.

additional_info.txt` Template for miscellaneous dataset information.

attributes_*.txt Template(s) for data table attribute metadata, populated with some automatically extracted metadata content from the data tables. A tab delimited table.

bounding_boxes.txt Template for the dataset geographic bounding coordinates. A tab delimited table.

custom_units.txt Template for defining non-standard units used in data tables. A tab delimited table.

intellectual_rights.txt The selected intellectual rights license.

keywords.txt Template for dataset keywords. A tab delimited table.

methods.txt Template for dataset methods.

personnel.txt Template for dataset personnel and funding metadata. A tab delimited table.

If using x, then the templates are added to /x/templates.

Details

Existing templates will not be overwritten subsequent calls to import_templates().