Import core metadata and table attribute templates

Import template files for storage of core metadata and table attributes. Some templates are populated with content detected by automated metadata extraction methods. The remainder is manually entered by the user. Instructions for editing the templates.

import_templates(
      path,
      data.path = path,
      license,
      data.table = NULL,
      write.file = TRUE,
      x = NULL
    )

Arguments

path
(character) Path to the directory where the templates will be written.

data.path
(character) Path to the directory where the data tables are stored.

license
(character) License under which the data will be released. Use https://creativecommons.org/publicdomain/zero/1.0/ or https://creativecommons.org/licenses/by/4.0/.

data.table
(character) Name of data table. If more than one, then supply as a vector of character strings (e.g. data.table = c('concentrations.csv', 'characteristics.csv')).

write.file
(logical; optional) Whether to write the templates to path.

x
(named list; optional) Alternative input to EMLassemblyline functions. Use template_arguments() to create x.

Value

abstract.txt Template for the dataset abstract.

additional_info.txt` Template for miscellaneous dataset information.

attributes_*.txt Template(s) for data table attribute metadata, populated with some automatically extracted metadata content from the data tables. A tab delimited table.

bounding_boxes.txt Template for the dataset geographic bounding coordinates. A tab delimited table.

custom_units.txt Template for defining non-standard units used in data tables. A tab delimited table.

intellectual_rights.txt The selected intellectual rights license.

keywords.txt Template for dataset keywords. A tab delimited table.

methods.txt Template for dataset methods.

personnel.txt Template for dataset personnel and funding metadata. A tab delimited table.

If using x, then the templates are added to /x/templates.

Details

Existing templates will not be overwritten subsequent calls to import_templates().

Examples

# Set working directory
setwd(tempdir())

# Create data package directory "pkg_250"
file.copy(
  from = system.file('/examples/pkg_250', package = 'EMLassemblyline'),
  to = tempdir(),
  recursive = TRUE
)

#> [1] TRUE

# View directory contents (NOTE: Directory is empty)
dir('./pkg_250/metadata_templates')

#> [1] "README.txt"

# Import templates (for data package with no tables)
import_templates(
  path = './pkg_250/metadata_templates',
  license = 'CC0'
)

#> Importing metadata templates

#> Warning: 'import_templates' is deprecated.
#> Use 'template_core_metadata and template_table_attributes' instead.
#> See help("Deprecated") and help("EMLassemblyline-deprecated").

#> Importing abstract.txt.

#> Importing additional_info.txt.

#> Importing bounding_boxes.txt.

#> Importing custom_units.txt.

#> Importing intellectual_rights.txt.

#> Importing keywords.txt.

#> Importing methods.txt.

#> Importing personnel.txt.

#> Done.

# View directory contents (NOTE: Templates exist)
dir('./pkg_250/metadata_templates')

#> [1] "abstract.txt"            "additional_info.txt"    
#> [3] "bounding_boxes.txt"      "custom_units.txt"       
#> [5] "intellectual_rights.txt" "keywords.txt"           
#> [7] "methods.txt"             "personnel.txt"          
#> [9] "README.txt"             

# Rerunning import_templates() does not overwrite files
import_templates(
  path = './pkg_250/metadata_templates',
  license = 'CC0'
)

#> Importing metadata templates

#> Warning: 'import_templates' is deprecated.
#> Use 'template_core_metadata and template_table_attributes' instead.
#> See help("Deprecated") and help("EMLassemblyline-deprecated").

#> abstract.txt already exists!

#> additional_info.txt already exists!

#> bounding_boxes.txt already exists!

#> custom_units.txt already exists!

#> intellectual_rights.txt already exists!

#> keywords.txt already exists!

#> methods.txt already exists!

#> personnel.txt already exists!

#> Done.

# Remove templates from directory
file.remove(
  list.files('./pkg_250/metadata_templates', full.names = TRUE)
)

#> [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE

# Import templates (for data package with tables)
import_templates(
  path = './pkg_250/metadata_templates',
  data.path = './pkg_250/data_objects',
  data.table = c(
    'decomp.csv',
    'nitrogen.csv'
  ),
  license = 'CC0'
)

#> Importing metadata templates

#> Warning: 'import_templates' is deprecated.
#> Use 'template_core_metadata and template_table_attributes' instead.
#> See help("Deprecated") and help("EMLassemblyline-deprecated").

#> Importing abstract.txt.

#> Importing additional_info.txt.

#> Importing bounding_boxes.txt.

#> Importing custom_units.txt.

#> Importing intellectual_rights.txt.

#> Importing keywords.txt.

#> Importing methods.txt.

#> Importing personnel.txt.

#> Importing attributes_decomp.txt.

#> Importing attributes_nitrogen.txt.

#> Done.

# View directory contents (NOTE: Templates, including attributes*_.txt, exist)
dir('./pkg_250/metadata_templates')

#>  [1] "abstract.txt"            "additional_info.txt"    
#>  [3] "attributes_decomp.txt"   "attributes_nitrogen.txt"
#>  [5] "bounding_boxes.txt"      "custom_units.txt"       
#>  [7] "intellectual_rights.txt" "keywords.txt"           
#>  [9] "methods.txt"             "personnel.txt"          

# Clean up
unlink(
  './pkg_250',
  recursive = TRUE
)