Render templates into EML metadata

Render the content of metadata templates into the EML, then validate the EML and write to file.

make_eml(
      path,
      data.path = path,
      eml.path = path,
      dataset.title,
      temporal.coverage,
      geographic.description,
      geographic.coordinates,
      maintenance.description,
      data.table = NULL,
      data.table.description = NULL,
      data.table.quote.character = NULL,
      other.entity = NULL,
      other.entity.description = NULL,
      data.url = NULL,
      provenance = NULL,
      user.id = NULL,
      user.domain = NULL,
      package.id = NULL,
      write.file = TRUE,
      return.obj = FALSE,
      x = NULL
    )

Arguments

path
(character) Path to where the template(s) are stored.

data.path
(character) Path to where the data files are stored.

eml.path
(character) Path to where the EML will be written.

dataset.title
(character) Dataset title. Be descriptive (more than 5 words) and consider including the temporal coverage (e.g. "GLEON: Long term lake chloride concentrations from North America and Europe: 1940-2016").

temporal.coverage
(character) Beginning and ending dates of the dataset as a vector of character strings in the format `YYYY-MM-DD`.

geographic.description
(character) Geographic description. Don't use this argument if geographic coordinates in the `bounding_boxes.txt` template.

geographic.coordinates
(character) Geographic coordinates delineating the bounding area or point of a dataset, in decimal degrees. Values must be listed in this order: North, East, South, West. Longitudes West of the prime meridian and latitudes South of the equator are negative. If representing a point, repeat the latitude for North and South, and repeat the longitude for East and West (e.g. `geographic.coordinates = c('28.38', '-119.95', '28.38', '-119.95)`).

maintenance.description
(character) Indicator of whether data collection is `ongoing` or `completed`.

data.table
(character) Data table name(s). If more than one, then supply as a vector of character strings (e.g. `data.table = c("concentrations.csv", "characteristics.csv")`).

data.table.description
(character) Data table description(s). Brief description of `data.table`. If more than one, then supply as a vector of character strings in the same order as listed in `data.table`.

data.table.quote.character
(character) Quote character(s) used in `data.table`. If more than one, then supply as a vector of character strings in the same order as listed in `data.table`. This argument is required only if your data contain quotations. If the quote character is a quotation, then enter `"\""`. If the quote character is an apostrophe, then enter `"\'"`.

other.entity
(character) Name(s) of `other.entity`(s) in this dataset. Use `other.entity` for all non-`data.table` files. `other.entity`(s) should be stored at `data.path`.

other.entity.description
(character) Description(s) of `other.entity`(s). If more than one, then supply as a vector of descriptions in the same order as listed in `other.entity`.

data.url
(character) The URL of where your data can be downloaded by a data repository. This argument is not required, if the data will be manually uploaded to a data repository.

provenance
(character) EDI Data Repository Data package ID(s) corresponding to parent datasets from which this dataset was created (e.g. `knb-lter-cap.46.3`).

user.id
(character) ID(s) of data repository user account(s). If more than one, supply as a vector of character strings. Contact EDI (info@environmentaldatainitiative.org) to obtain one. This argument is not required.

user.domain
(character) Domain of the `user.id`(s). Valid options for EDI are `LTER` and `EDI`. If more than one, supply as a vector of character strings in the same order as corresponding `user.id`(s). This argument is not required.

package.id
(character) EDI Data Repository Data package ID for this dataset. A missing package ID defaults to `edi.101.1`.

write.file
(logical) Write file to `eml.path`.

return.obj
(logical) Return the EML as an R object of class `EML object`.

x
(named list) Alternative input/output to `EMLassemblyline` functions. Use template_arguments to create `x`.

Value

`Status messages` describing the EML creation status

`Validation result` indicating whether the EML is schema valid

`EML file` written to `path` and in the form `package.id.xml`

`EML object` when `return.obj = TRUE`

Details

If validation fails, open the EML document in a .xml editor to identify the source of error. Often the error is small and quickly resolved with the aid of an editors schema validator. If the issue can't be resolved then forward the `EML` to info@environmentaldatainitiative.org for support.