Make EML metadata

Render the contents of metadata templates into EML, validate, and write to file.

make_eml(
      path,
      data.path = path,
      eml.path = path,
      dataset.title,
      temporal.coverage,
      geographic.description,
      geographic.coordinates,
      maintenance.description,
      data.table = NULL,
      data.table.name = data.table,
      data.table.description = NULL,
      data.table.quote.character = NULL,
      other.entity = NULL,
      other.entity.name = other.entity,
      other.entity.description = NULL,
      data.url = NULL,
      provenance = NULL,
      user.id = NULL,
      user.domain = NULL,
      package.id = NULL,
      write.file = TRUE,
      return.obj = FALSE,
      x = NULL
    )

Arguments

path
(character) Path to the metadata template directory.

data.path
(character) Path to the data directory.

eml.path
(character) Path to the EML directory, where EML files are written.

dataset.title
(character) Title of the dataset.

temporal.coverage
(character) Beginning and ending dates of the dataset in the format "YYYY-MM-DD" (e.g. temporal.coverage = c('2012-05-01', '2014-11-30')).

geographic.description
(character) Description of datasets geographic extent. Don't use this argument if geographic coverage is supplied by geographic_coverage.txt.

geographic.coordinates
(character) Coordinates of datasets geographic extent. Coordinates are listed in this order: North, East, South, West (e.g. geographic.coordinates = c('28.38', '-119.95', '28.38', '-119.95')). Longitudes west of the prime meridian and latitudes south of the equator are negative. Don't use this argument if geographic coverage is supplied by geographic_coverage.txt.

maintenance.description
(character) Data collection status ("ongoing" or "complete").

data.table
(character; optional) Table file name. If more than one, then supply as a vector of character strings (e.g. data.table = c("nitrogen.csv", "decomp.csv")).

data.table.name
(character; optional) A short descriptive name for the table. Defaults to data.table. If more than one, then supply as a vector of character strings in the same order as listed in data.table.

data.table.description
(character; optional) Table description. If more than one, then supply as a vector of character strings in the same order as listed in data.table.

data.table.quote.character
(character; optional) Quote character used in data.table. If more than one, then supply as a vector of character strings in the same order as listed in data.table. If the quote character is a quotation, then enter "\"". If the quote character is an apostrophe, then enter "\'".

other.entity
(character; optional) Name of other.entity(s) in this dataset. Use other.entity for all non-data.table files. other.entity(s) should be stored at data.path. If more than one, then supply as a vector of character strings (e.g. other.entity = c('ancillary_data.zip', 'quality_control.R')).

other.entity.name
(character; optional) A short descriptive name for the other.entity. Defaults to other.entity. If more than one, then supply as a vector of character strings in the same order as listed in other.entity.

other.entity.description
(character; optional) Description(s) of other.entity(s). If more than one, then supply as a vector of descriptions in the same order as listed in other.entity.

data.url
(character; optional) The publicly accessible URL from which data.tables or other.entitys can be downloaded. This argument is not required, if the data will be uploaded manually to the data repository.

provenance
(character; optional) EDI Data Repository Data package ID(s) corresponding to parent datasets from which this dataset was created (e.g. knb-lter-cap.46.3).

user.id
(character; optional) ID(s) of data repository user account(s). If more than one, supply as a vector of character strings.

user.domain
(character; optional) Domain of the user.id(s). Valid options for EDI are "LTER" and "EDI". If more than one, supply as a vector of character strings in the same order as corresponding user.id(s).

package.id
(character; optional) Data Repository data package ID for this dataset. A missing package ID defaults to "edi.101.1".

write.file
(logical; optional) Whether to write the EML file.

return.obj
(logical; optional) Whether to return the EML as an R object of class EML object.

x
(named list; optional) Alternative input to make_eml(). Use template_arguments() to create x.

Value

EML file written to eml.path.

EML object when return.obj = TRUE.

Details

make_eml() reads the contents of metadata templates, auto-extracts additional metadata from the data entities, appends value added content (e.g. resolving keywords to controlled vocabularies), and adds all the metadata content to locations in the EML schema according with best practice recommendations of scientists, data managers, and data repositories. The EML is then validated against the schema and written to file.

Examples

# Initialize data package directory for make_eml()
file.copy(
  from = system.file('/examples/pkg_260', package = 'EMLassemblyline'),
  to = tempdir(),
  recursive = TRUE
)

#> [1] TRUE

# Set working directory
setwd(paste0(tempdir(), '/pkg_260'))

# Make EML (for data package with data tables)

make_eml(
  path = './metadata_templates',
  data.path = './data_objects',
  eml.path = './eml',
  dataset.title = 'Sphagnum and Vascular Plant Decomposition under Increasing Nitrogen Additions',
  temporal.coverage = c('2014-05-01', '2015-10-31'),
  maintenance.description = 'completed',
  data.table = c('decomp.csv', 'nitrogen.csv'),
  data.table.description = c('Decomposition data', 'Nitrogen data'),
  user.id = 'csmith',
  user.domain = 'EDI',
  package.id = 'edi.260.1'
)

#> Making EML ...

#> Warning: Defaulting argument 'data.table.name' to 'data.table'. Consider providing a brief descriptive name for your file(s).

#> Creating nodes ...

#> <eml>

#>   <access>

#>   <dataset>

#>     <title>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <associatedParty>

#>     <associatedParty>

#>     <associatedParty>

#>     <associatedParty>

#>     <pubDate>

#>     <abstract>

#>     <keywordSet>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keywordThesaurus>

#>    <additionalInfo>

#>     <intellectualRights>

#>     <coverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <temporalCoverage>

#>       <taxonomicCoverage>

#>     <maintenance>

#>     <contact>

#>     <contact>

#>     <contact>

#>     <methods>

#>     <project>

#>       <relatedProject>

#>       <relatedProject>

#>       <relatedProject>

#>       <relatedProject>

#>     <dataTable> (decomp.csv)

#>     <dataTable> (nitrogen.csv)

#>     <additionalMetadata>

#>   <additionalMetadata>

#> </eml>

#> Writing EML (edi.260.1.xml)

#> Error in ns_lookup(parent$doc, parent$node, parts[[1]]): No namespace with prefix `https` found

# View EML directory contents (NOTE: edi.260.1 exists)
dir('./eml')

#> [1] "README.txt"

# Make EML (for data package with other entities)

make_eml(
  path = './metadata_templates',
  data.path = './data_objects',
  eml.path = './eml',
  dataset.title = 'Sphagnum and Vascular Plant Decomposition under Increasing Nitrogen Additions',
  temporal.coverage = c('2014-05-01', '2015-10-31'),
  maintenance.description = 'completed',
  other.entity = c('ancillary_data.zip', 'processing_and_analysis.R'),
  other.entity.description = c('Ancillary data', 'Data processing and analysis script'),
  user.id = 'csmith',
  user.domain = 'EDI',
  package.id = 'edi.260.2'
)

#> Making EML ...

#> Warning: Defaulting argument 'other.entity.name' to 'other.entity'. Consider providing a brief descriptive name for your file(s).

#> Creating nodes ...

#> <eml>

#>   <access>

#>   <dataset>

#>     <title>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <associatedParty>

#>     <associatedParty>

#>     <associatedParty>

#>     <associatedParty>

#>     <pubDate>

#>     <abstract>

#>     <keywordSet>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keywordThesaurus>

#>    <additionalInfo>

#>     <intellectualRights>

#>     <coverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <temporalCoverage>

#>       <taxonomicCoverage>

#>     <maintenance>

#>     <contact>

#>     <contact>

#>     <contact>

#>     <methods>

#>     <project>

#>       <relatedProject>

#>       <relatedProject>

#>       <relatedProject>

#>       <relatedProject>

#>     <otherEntity> (ancillary_data.zip)

#>     <otherEntity> (processing_and_analysis.R)

#> </eml>

#> Writing EML (edi.260.2.xml)

#> Validating EML

#> EML validaton failed. See warnings for details.

#> Done.

# View EML directory contents (NOTE: edi.260.2 exists)
dir('./eml')

#> [1] "edi.260.2.xml" "README.txt"   

# Make EML (for data package with data tables and other entities)

make_eml(
  path = './metadata_templates',
  data.path = './data_objects',
  eml.path = './eml',
  dataset.title = 'Sphagnum and Vascular Plant Decomposition under Increasing Nitrogen Additions',
  temporal.coverage = c('2014-05-01', '2015-10-31'),
  maintenance.description = 'completed',
  data.table = c('decomp.csv', 'nitrogen.csv'),
  data.table.description = c('Decomposition data', 'Nitrogen data'),
  other.entity = c('ancillary_data.zip', 'processing_and_analysis.R'),
  other.entity.description = c('Ancillary data', 'Data processing and analysis script'),
  user.id = 'csmith',
  user.domain = 'EDI',
  package.id = 'edi.260.3'
)

#> Making EML ...

#> Warning: Defaulting argument 'data.table.name' to 'data.table'. Consider providing a brief descriptive name for your file(s).

#> Warning: Defaulting argument 'other.entity.name' to 'other.entity'. Consider providing a brief descriptive name for your file(s).

#> Creating nodes ...

#> <eml>

#>   <access>

#>   <dataset>

#>     <title>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <creator>

#>     <associatedParty>

#>     <associatedParty>

#>     <associatedParty>

#>     <associatedParty>

#>     <pubDate>

#>     <abstract>

#>     <keywordSet>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keyword>

#>     <keywordThesaurus>

#>    <additionalInfo>

#>     <intellectualRights>

#>     <coverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <geographicCoverage>

#>       <temporalCoverage>

#>       <taxonomicCoverage>

#>     <maintenance>

#>     <contact>

#>     <contact>

#>     <contact>

#>     <methods>

#>     <project>

#>       <relatedProject>

#>       <relatedProject>

#>       <relatedProject>

#>       <relatedProject>

#>     <dataTable> (decomp.csv)

#>     <dataTable> (nitrogen.csv)

#>     <additionalMetadata>

#>     <otherEntity> (ancillary_data.zip)

#>     <otherEntity> (processing_and_analysis.R)

#>   <additionalMetadata>

#> </eml>

#> Writing EML (edi.260.3.xml)

#> Validating EML

#> EML validaton failed. See warnings for details.

#> Done.

# View EML directory contents (NOTE: edi.260.3 exists)
dir('./eml')

#> [1] "edi.260.2.xml" "edi.260.3.xml" "README.txt"   

# Clean up
unlink('.', recursive = TRUE)