Read template and data files into a list structure

Read metadata templates, data files, and function arguments into an R list structure as an alternative input to EMLassemblyline functions (i.e. rather than supplying the files themselves).

This function enables programatic interfacing of upstream sources with EMLassemblyline (e.g. a metabase or database). See below for valid list structure.

read_files(
      path,
      data.path = path,
      data.table = NULL,
      other.entity = NULL,
      sep = NULL
    )

Arguments

path
(character) Path to metadata templates.

data.path
(character) Path to data tables and other entities.

data.table
(character) Data table name(s). If more than one, then supply as a vector of character strings (e.g. data.table = c("concentrations.csv", "characteristics.csv")).

other.entity
(character) Name(s) of other.entity(s) in this dataset. Use other.entity for all non-data.table files. other.entity(s) should be stored at data.path.

sep
(character) Delimiter of data.table. Use this argument if read_files() fails to automatically identify data.table field delimiters.

Value

(named list) Named list with this internal structure.:

template Metadata template files

abstract.txt Abstract template

content EML::TextType containing abstract, NA otherwise

path Path to file, NA otherwise

additional_info.txt Additional information template

content EML::TextType containing additional info, NA otherwise

path Path to file, NA otherwise

attributes_name.txt Attributes template where "name" is name of data.table described by the attributes. Don't include if missing.

content Data frame of attributes template, NA otherwise

path Path to file, NA otherwise

bounding_boxes.txt Bounding boxes template

content Data frame of bounding boxes template, NA otherwise

path Path to file, NA otherwise

catvars_name.txt Categorical variables template where name is name of data.table described by the catvars. Don't include if missing.

content Data frame of categorical variables template, NA otherwise

path Path to file, NA otherwise

custom_units.txt Custom units template

content Data frame of custom units template, NA otherwise

path Path to file, NA otherwise

geographic_coverage.txt Geographic coverage template

content Data frame of geographic coverage template, NA otherwise

path Path to file, NA otherwise

intellectual_rights.txt Intellectual rights template

content EML::TextType containing intellectual rights, NA otherwise

path Path to file, NA otherwise

keywords.txt Keywords template

content Data frame of keywords template, NA otherwise

path Path to file, NA otherwise

methods.txt Methods template

content EML::Methods containing methods, NA otherwise

path Path to file, NA otherwise

personnel.txt Personnel template

content Data frame of personnel template, NA otherwise

path Path to file, NA otherwise

taxonomicCoverage.xml Taxonomic coverage EML element

content EML::taxonomicCoverage, NA otherwise

path Path to file, NA otherwise

data.table Data tables

data.table.name.ext Name of data table including file extension, NULL otherwise

content Data frame of data table

path Path to data table (required if data.table exists)

other.entity Other entities

other.entity.name.ext Name of other entity including file extension, NULL otherwise

content Set to NA

path Path to other entity (required if data.table exists)

argument A list of arguments to be passed to make_eml(). Other EMLassemblyline functions don't yet accept this input.

data.path (character) Path to where the data files are stored. This path must be valid when passing the list structure to make_eml() because this is where checksum values are calculated.

data.table (character) Data table name(s). If more than one, then supply as a vector of character strings (e.g. data.table = c("concentrations.csv", "characteristics.csv")).

data.table.description (character) Data table description(s). Brief description of data.table. If more than one, then supply as a vector of character strings in the same order as listed in data.table.

data.table.quote.character (character) Quote character(s) used in data.table. If more than one, then supply as a vector of character strings in the same order as listed in data.table. This argument is required only if your data contain quotations. If the quote character is a quotation, then enter "\"". If the quote character is an apostrophe, then enter "\'".

data.url (character) The URL of where your data can be downloaded by a data repository. This argument is not required, if the data will be manually uploaded to a data repository.

dataset.title (character) Dataset title. Be descriptive (more than 5 words) and consider including the temporal coverage (e.g. "GLEON: Long term lake chloride concentrations from North America and Europe: 1940-2016").

eml.path (character) Path to where the EML will be written.

geographic.coordinates (character) Geographic coordinates delineating the bounding area or point of a dataset, in decimal degrees. This argument is not required if using bounding_boxes.txt. Values must be listed in this order: North, East, South, West. Longitudes West of the prime meridian and latitudes South of the equator are negative. If representing a point, repeat the latitude for North and South, and repeat the longitude for East and West (e.g. geographic.coordinates = c('28.38', '-119.95', '28.38', '-119.95')).

geographic.description (character) Geographic description. Don't use this argument if geographic coordinates in the bounding_boxes.txt template.

maintenance.description (character) Indicator of whether data collection is ongoing or completed.

other.entity (character) Name(s) of other.entity(s) in this dataset. Use other.entity for all non-data.table files. other.entity(s) should be stored at data.path.

other.entity.description (character) Description(s) of other.entity(s). If more than one, then supply as a vector of descriptions in the same order as listed in other.entity.

package.id (character) EDI Data Repository Data package ID for this dataset. A missing package ID defaults to edi.101.1.

path (character) Path to where the template(s) will be stored.

provenance (character) EDI Data Repository Data package ID(s) corresponding to parent datasets from which this dataset was created (e.g. knb-lter-cap.46.3).

return.obj (logical) Return the EML as an R object of class EML object.

temporal.coverage (character) Beginning and ending dates of the dataset as a vector of character strings in the format YYYY-MM-DD.

user.domain (character) Domain of the user.id(s). Valid options for EDI are LTER and EDI. If more than one, supply as a vector of character strings in the same order as corresponding user.id(s). This argument is not required.

user.id (character) ID(s) of data repository user account(s). If more than one, supply as a vector of character strings. Contact EDI (info@environmentaldatainitiative.org) to obtain one. This argument is not required.

write.file (logical) Write file to eml.path.

Note

Paths to data tables and other entities must be valid when passing the list structure to make_eml(). Paths to other content is optional.