Initialize list of arguments for EMLassemblyline functions

Initialize the list of arguments for EMLassemblyline functions including x, an argument containing metadata template and data file content. This approach enables programatic interfacing of upstream metadata sources (e.g. LTER-core-metabase) to EMLassemblyline functions.

template_arguments(
      path = NULL,
      data.path = NULL,
      data.table = NULL,
      other.entity = NULL,
      sep = NULL
    )

Arguments

path
(character) Path to metadata templates.

data.path
(character) Path to data tables and other entities.

data.table
(character) Data table name. If more than one, then supply as a vector of character strings (e.g. data.table = c('concentrations.csv', 'characteristics.csv')).

other.entity
(character) Other entity name. If more than one, then supply as a vector of character strings (e.g. other.entity = c('ancillary_data.zip', 'quality_control.R')).

sep
(character) Data table field delimiter. Use this argument if template_arguments fails to automatically identify field delimiters of data.table.

Value

(named list) A list of all EMLassemblyline arguments with this structure:

data.path (character) Path to where the data files are stored.

data.table (character) Data table name. If more than one, supply as a vector of character strings (e.g. data.table = c('concentrations.csv', 'characteristics.csv')).

data.table.description (character) Data table description. Brief description of data tables. If more than one data table, then supply as a vector of character strings in the same order as listed in data.table.

data.table.quote.character (character) Quote character used in data tables. If more than one, then supply as a vector of character strings in the same order as listed in data.table. This argument is required only if data.table contain quotations. If the quote character is a quotation, then enter "\"". If the quote character is an apostrophe, then enter "\'".

data.url (character) A publicly accessible URL to data tables and other entities of this dataset.

dataset.title (character) Dataset title.

dir.name (character) Name of parent directory. Subdirectories have predefined names.

eml.path (character) Path to where EML will be written.

geographic.coordinates (character) Geographic coordinates delineating the bounding area or point of a dataset, in decimal degrees. This argument is not required if using bounding_boxes.txt. Values must be listed in this order: North, East, South, West. Longitudes West of the prime meridian and latitudes South of the equator are negative. If representing a point, repeat the latitude for North and South, and repeat the longitude for East and West (e.g. geographic.coordinates = c('28.38', '-119.95', '28.38', '-119.95')).

geographic.description (character) Geographic description. Don't use this argument if using the bounding_boxes.txt template.

lat.col (character) Name of latitude column.

license (character) License under which the data will be released.

lon.col (character) Name of longitude column.

maintenance.description (character) Indicator of whether data collection is 'ongoing' or 'completed'.

other.entity (character) Other entity name. If more than one, then supply as a vector of character strings. Other entities should be located at data.path.

other.entity.description (character) Other entity description. If more than one, then supply as a vector of descriptions in the same order as listed in other.entity.

package.id (character) Data package identifier. Missing package.id is assigned a default value of 'edi.101.1'. Any data repository package ID may be used.

path (character) Path to where template files are stored.

provenance (character) EDI Data Repository Data package ID(s) corresponding to parent datasets from which this dataset was created (e.g. 'knb-lter-cap.46.3'). If more than one, then supply as a vector of character strings.

return.obj (logical) Return the EML as an R object of class EML object.

sep (character) Field delimiter to be used when reading data tables.

site.col (character) Name of site column, where site is the name of the location specified by lat.col and lon.col.

taxa.authority (integer) An ordered numeric vector of ID's corresponding to data sources (i.e. taxonomic authorities) you'd like to query, in the order of decreasing preference. Run view_taxa_authorities to see supported data sources. Columns "resolve_sci_taxa", and "resolve_comm_taxa" correspond to scientific and common searches.

taxa.col (character) Name of column in taxa.tablecontaining taxonomic names.

taxa.name.type (character) Taxonomic name type. Can be: scientific, common, or both.

taxa.table (character) Name of data table containing taxa.col.

temporal.coverage (character) Beginning and ending dates of the dataset as a vector of character strings in the format YYYY-MM-DD.

user.domain (character) Domain of the user.id. If more than one, then supply as a vector of character strings in the same order corresponding to listed under user.id. Valid options for EDI are 'LTER' and 'EDI'. Other data repository IDs may be used. NULL user.domain defaults to 'someuserdomain'.

user.id (character) ID of data repository user account. If more than one, supply as a vector of character strings. NULL user.id defaults to 'someuserid'.

write.file (logical) Write file to eml.path.

x (named list) List of metadata template, data tables, and other entities.

template - Metadata template files

abstract.txt Abstract template

content EML103::TextType containing abstract, NA otherwise

additional_info.txt Additional information template

content EML103::TextType containing additional info, NA otherwise

attributes_name.txt Attributes template where "name" is the corresponding data.table.

content Data frame of attributes template, NA otherwise.

bounding_boxes.txt Bounding boxes template

content Data frame of bounding boxes template, NA otherwise.

catvars_name.txt Categorical variables template where "name" is the corresponding data.table.

content Data frame of categorical variables template, NA otherwise.

custom_units.txt Custom units template

content Data frame of custom units template, NA otherwise.

geographic_coverage.txt Geographic coverage template

content Data frame of geographic coverage template, NA otherwise.

intellectual_rights.txt Intellectual rights template

content EML103::TextType containing intellectual rights template, NA otherwise.

keywords.txt Keywords template

content Data frame of keywords template, NA otherwise.

methods.txt Methods template

content EML103::Methods containing methods, NA otherwise.

personnel.txt Personnel template

content Data frame of personnel template, NA otherwise.

taxonomicCoverage.xml Taxonomic coverage EML element

content EML103::taxonomicCoverage, NA otherwise

data.table Data tables

data.table.name.ext Name of data table including file extension, NULL otherwise.

content Data frame of data table.

other.entity Other entities

other.entity.name.ext Name of other entity including file extension, NULL otherwise.

content NA

Note

The full named list structure is returned irrespective of NULL inputs, so content can be added downstream. If valid inputs are supplied, then the respective list structures are populated. Pass the output of this function to make_template_core to initilize (or add) core template data frames, text objects, or XML objects.