Create core metadata templates

Use this function to create the core metadata templates required by all data packages. Most templates require user supplied information. Instructions for editing these templates.

template_core_metadata(
      path,
      license,
      file.type = '.txt',
      write.file = TRUE,
      x = NULL
    )

Arguments

path
(character) Path to the metadata template directory.

license
(character) License to publicly release the data package under. Use "CC0" or "CCBY".

file.type
(character; optional) File type for abstract, methods, and additional info metadata templates. Can be: '.txt' (plain text), '.docx' (MS Word), or '.md' (markdown).

write.file
(logical; optional) Whether to write the template files.

x
(named list; optional) Alternative input to template_core_metadatta(). Use template_arguments() to create x.

Value

abstract The abstract template.

additional_info The template for miscellaneous information.

intellectual_rights.txt The intellectual rights license with the text of CC0 or CCBY.

keywords.txt The tab delimited keywords template.

methods The methods template.

personnel.txt The tab delimited personnel template for information on persons and funding involved in the creation of the data package.

These files is written to path unless using x, in which case the templates are added to /x/templates.

Details

Existing templates will not be overwritten by template_core_metadata().

Examples

# Initialize data package directory for template_core_metadata()
file.copy(
  from = system.file('/examples/pkg_250', package = 'EMLassemblyline'),
  to = tempdir(),
  recursive = TRUE
)

#> [1] TRUE

# Set working directory
setwd(paste0(tempdir(), '/pkg_250'))

# View directory contents (NOTE: Directory is empty)
dir('./metadata_templates')

#> [1] "README.txt"

# Template core metadata
template_core_metadata(
  path = './metadata_templates',
  license = 'CC0'
)

#> Templating core metadata ...

#> abstract.txt

#> additional_info.txt

#> intellectual_rights.txt

#> keywords.txt.

#> methods.txt

#> personnel.txt

#> Done.

# View directory contents (NOTE: Templates exist)
dir('./metadata_templates')

#> [1] "abstract.txt"            "additional_info.txt"    
#> [3] "intellectual_rights.txt" "keywords.txt"           
#> [5] "methods.txt"             "personnel.txt"          
#> [7] "README.txt"             

# Rerunning template_core_metadata() does not overwrite files
template_core_metadata(
  path = './metadata_templates',
  license = 'CC0'
)

#> Templating core metadata ...

#> abstract.txt already exists!

#> additional_info.txt already exists!

#> intellectual_rights.txt already exists!

#> keywords.txt already exists!

#> methods.txt already exists!

#> personnel.txt already exists!

#> Done.

# Use docx files for abstract, methods, and additional_info
# First remove abstract, methods, and additional_info templates
file.remove(
  c(
    './metadata_templates/abstract.txt',
    './metadata_templates/methods.txt',
    './metadata_templates/additional_info.txt'
  )
)

#> [1] TRUE TRUE TRUE

# Create docx templates
template_core_metadata(
  path = './metadata_templates',
  license = 'CC0',
  file.type = '.docx'
)

#> Templating core metadata ...

#> abstract.docx

#> additional_info.docx

#> intellectual_rights.txt already exists!

#> keywords.txt already exists!

#> methods.docx

#> personnel.txt already exists!

#> Done.

# Use md files for abstract, methods, and additional_info
# First remove abstract, methods, and additional_info templates
file.remove(
  c(
    './metadata_templates/abstract.docx',
    './metadata_templates/methods.docx',
    './metadata_templates/additional_info.docx'
  )
)

#> [1] TRUE TRUE TRUE

# Create md templates
template_core_metadata(
  path = './metadata_templates',
  license = 'CC0',
  file.type = '.md'
)

#> Templating core metadata ...

#> abstract.md

#> additional_info.md

#> intellectual_rights.txt already exists!

#> keywords.txt already exists!

#> methods.md

#> personnel.txt already exists!

#> Done.

# Clean up
unlink('.', recursive = TRUE)