Initialize directories for templates, data objects, and EML

Create a directory structure for data package contents and EMLassemblyline files. Any organization scheme is supported. This is one recommendation.

template_directories(path, dir.name)

Arguments

path
(character) Path to where the data package directory will be created.

dir.name
(character) Name of the directory.

Value

A directory with the following structure and contents:

name Name supplied via dir.name

data_objects Directory for data and other objects to be packaged.

metadata_templates Directory for EMLassemblyline template files.

eml Directory for EML files created by EMLassemblyline.

run_EMLassemblyline.R An empty R file for scripting an EMLassemblyline workflow.

Details

Existing directories named with dir.name at path will not be overwritten.

Examples

if (FALSE) {
# Set working directory
setwd("/Users/me/Documents/data_packages")

# For data package 260
template_directories(
  path = ".",
  dir.name =  "pkg_260")
}