Create taxonomic coverage template

Use this function to extract the unique taxa of a data table and try to resolve (match) to a taxonomic authority system (e.g. ITIS) and return for user supplied inputs if necessary. This information is later used to list the full hierarchical rank names in the metadata. Instructions for editing this template.

template_taxonomic_coverage(
      path,
      data.path = path,
      taxa.table,
      taxa.col,
      taxa.name.type,
      taxa.authority,
      write.file = TRUE,
      x = NULL
    )

Arguments

path
(character) Path to the metadata template directory.

data.path
(character) Path to the data directory.

taxa.table
(character) Table name containing taxa.col. If inputting more than one table, then supply as a list (e.g. c('table.1', 'table.2')).

taxa.col
(character) Column name containing taxa names. Species binomials are accepted. If inputting more than one table, then supply as a list c('taxa.col.table.1', 'taxa.col.table.2').

taxa.name.type
(character) Taxa name type. Can be: scientific, common, or both.

taxa.authority
(integer) Taxonomic authorities to resolve to as an ordered numeric vector of ID's decreasing in preference. See the list of supported authorities with taxonomyCleanr::view_taxa_authorities(). Columns "resolve_sci_taxa", and "resolve_comm_taxa" list authorites supporting scientific and common searches, respectively.

write.file
(logical; optional) Whether to write the template file.

x
(named list; optional) Alternative input to template_taxonomic_coverage(). Use template_arguments() to create x.

Value

taxonomic_coverage.txt The tab delimited taxonomic coverage template. This file is written to path unless using x, in which case the template is added to /x/templates/taxonomic_coverage.txt. Non-resolved taxa result NA.

Details

template_taxonomic_coverage() searches the most preferred taxonomic authority for all unique taxa listed in taxa.col returning the authority name and corresponding taxa identifier for direct matches (no fuzzy searching), then the next most preferred taxonomic authority is search for taxa that have not yet been resolved. This process repeats for subsequently listed authorities. "NA" is returned when an authority match is not made.

When taxonomic_coverage.txt is passed to make_eml(), the authority information is used to get the hierarchical rank names of resolved taxa and rendered into the "taxonomicCoverage" element of EML.

Existing taxonomic_coverage.txt will not be overwritten by subsequent calls to template_taxonomic_coverage().

Examples

# Initialize data package directory for template_taxonomic_coverage()
file.copy(
  from = system.file('/examples/pkg_255', package = 'EMLassemblyline'),
  to = tempdir(),
  recursive = TRUE
)

#> [1] TRUE

# Set working directory
setwd(paste0(tempdir(), '/pkg_255'))

# View directory contents (NOTE: taxonomic_coverage.txt doesn't exist)
dir('./metadata_templates')

#>  [1] "abstract.txt"            "additional_info.txt"    
#>  [3] "attributes_decomp.txt"   "attributes_nitrogen.txt"
#>  [5] "catvars_decomp.txt"      "catvars_nitrogen.txt"   
#>  [7] "custom_units.txt"        "geographic_coverage.txt"
#>  [9] "intellectual_rights.txt" "keywords.txt"           
#> [11] "methods.txt"             "personnel.txt"          

# Template taxonomic coverage
template_taxonomic_coverage(
  path = './metadata_templates',
  data.path = './data_objects',
  taxa.table = 'decomp.csv',
  taxa.col = 'taxa',
  taxa.authority = c(3,11),
  taxa.name.type = 'scientific'
)

#> Templating taxonomic coverage ...

#> Searching ITIS for "Mosses"

#> Searching GBIF Backbone Taxonomy for "Mosses"

#> Searching ITIS for "Unsorted biomass"

#> Searching GBIF Backbone Taxonomy for "Unsorted biomass"

#> Searching ITIS for "Achillea millefolium"

#> Searching ITIS for "Bouteloua gracilis"

#> Searching ITIS for "Cyperus"

#> Searching ITIS for "Koeleria cristata"

#> Searching ITIS for "Lespedeza capitata"

#> Searching ITIS for "Liatris aspera"

#> taxonomic_coverage.txt

#> Done.

# View directory contents (NOTE: taxonomic_coverage.txt exists)
dir('./metadata_templates')

#>  [1] "abstract.txt"            "additional_info.txt"    
#>  [3] "attributes_decomp.txt"   "attributes_nitrogen.txt"
#>  [5] "catvars_decomp.txt"      "catvars_nitrogen.txt"   
#>  [7] "custom_units.txt"        "geographic_coverage.txt"
#>  [9] "intellectual_rights.txt" "keywords.txt"           
#> [11] "methods.txt"             "personnel.txt"          
#> [13] "taxonomic_coverage.txt" 

# Rerunning template_taxonomic_coverage() does not overwrite file
template_taxonomic_coverage(
  path = './metadata_templates',
  data.path = './data_objects',
  taxa.table = 'decomp.csv',
  taxa.col = 'taxa',
  taxa.authority = c(3,11),
  taxa.name.type = 'scientific'
)

#> Templating taxonomic coverage ...

#> taxonomic_coverage.txt already exists!

#> Done.

# Clean up
unlink('.', recursive = TRUE)